% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-utils.R
\name{active_scenario}
\alias{active_scenario}
\title{Get the main directories of the currently active scenario folder.}
\usage{
active_scenario()
}
\value{
a list of 3 elements: 'scenario', 'inputs' and 'outputs' directories.
}
\description{
Get the main directories of the currently active scenario folder.
}
