% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity-functions.R
\name{add_history}
\alias{add_history}
\title{Add history to entity}
\usage{
add_history(entity, ids, event, time = .get_sim_time(), id_col_as_list = FALSE)
}
\arguments{
\item{entity}{an \link{Entity} or its inheritances object.}

\item{ids}{a integer vector that contains ids of entities to record the event}

\item{event}{a character value that denotes the event that is being recorded}

\item{time}{a integer value that represent the time at which the event occured
for the entities.}

\item{id_col_as_list}{a logical value which indicates whether the id column
should be stored as a list column. This significantly reduces the memory
footprint of the history data with a higher cost for basic data manipulation
since the data must be unnested to filter by entity ids.}
}
\description{
Recording past events of entities allow them to have memory of their past
experiences and actions. This can be later use, if requires, in future decision
making processes. For example, the number of past fertility events may be a
strong predictor of the next fertility for female individuals.
}
\examples{

create_toy_population()
add_history(pop$get("Individual"), ids = c(1:100), event = "test_event", time = 1)
pop$get("Individual")$get_data("history")
}
