% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{get_log}
\alias{get_log}
\alias{get_log.Generic}
\alias{get_log.World}
\alias{get_log.Container}
\alias{add_log}
\alias{add_log.World}
\alias{add_log.Container}
\alias{add_log.Entity}
\title{Add log to a Container or Entity object}
\usage{
get_log(x)

\method{get_log}{Generic}(x)

\method{get_log}{World}(x)

\method{get_log}{Container}(x)

add_log(x, ...)

\method{add_log}{World}(x, entity = NULL, desc, value, tag = NA_character_, ...)

\method{add_log}{Container}(x, desc, value, tag = NA_character_, ...)

\method{add_log}{Entity}(x, desc, value, tag = NA_character_, ...)
}
\arguments{
\item{x}{(\link{Container}|\link{Entity})\cr
an instance of \link[R6:R6Class]{R6::R6Class} to add a log entry to.}

\item{...}{dots}

\item{entity}{(\code{character})\cr
an \link{Entity} classname.}

\item{desc}{(\code{character(1)})\cr
a log description.}

\item{value}{(\code{any})\cr
an object to be stored. You can even store an \link{xtabs} object, if you wish.}

\item{tag}{(\code{character(1)})\cr
a tag, default as \code{NA_character}.}
}
\value{
\code{add_log} invinsibly returns \code{x}, while \code{get_log} returns \link[data.table:data.table]{data.table::data.table}
with the following structure:\preformatted{data.table(time = integer(),
           created_timestamp = integer(),
           class = character(),
           tag = character(),
           desc = character(),
           value = list())
}
}
\description{
This is an S3 method for adding logs to a \link{Container} or \link{Entity} object in
a microsimulation pipeline. This method passes all the arguments to the \code{log()}
method of the object in \code{x}.
}
\examples{

create_toy_world()
add_log(world, entity = "Individual", desc = "count:individuals", value = world$entities$Individual$n())
get_log(world)
}
\concept{logging}
