% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-utils.R
\name{assign_reference}
\alias{assign_reference}
\title{Assign reference}
\usage{
assign_reference(x, object_class)
}
\arguments{
\item{x}{a \link{World} object.}

\item{object_class}{a character or a \link[R6:R6Class]{R6::R6Class} generator.}
}
\value{
an \link[R6:R6Class]{R6::R6Class} object
}
\description{
A warper function of \link{World}'s get() method.
}
\examples{

create_toy_world()
# both calls are equivalent
Pop <- assign_reference(world, "Population")
Pop <- assign_reference(world, Population)
}
