% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{bioc_packages}
\alias{bioc_packages}
\title{Bioconductor Packages}
\usage{
bioc_packages()
}
\value{
Character vector
}
\description{
Simply return a vector with names of bioconductor packages
}
\details{
The list of packages includes:

Installer: BiocInstaller

Genomic Arithmetic: Biostrings, IRanges, GenomicRanges, rtracklayer

Differential Expression: limma, edgeR, DESeq2

Performance: parallel
}
\author{
Malte Thodberg
}

