% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{Building}
\alias{Building}
\title{Building class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Create Building class, extended Agent class.
}
\section{Construction}{
\preformatted{x <- Entity$new(databackend, .data, id_col)
}

Stores \code{.data} as a DataBackend object inside the object's list of data (\code{private$.data})
and registers the \code{id_col} (\code{private$.id_col}).
\itemize{
\item \code{databackend} :: a \link[R6:R6Class]{R6::R6Class} generator\cr
A \link[R6:R6Class]{R6::R6Class} generator that inherits from \code{DataBackend}.
\item \code{.data} :: \code{data.frame()}\cr
A object that inherits from \code{data.frame}.
\item \code{id_col} :: \code{character(1)}\cr
The id column of \code{.data}.
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{is_occupied(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is occupied
\item \code{is_vacant(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is vacant.
}
}

\section{Super classes}{
\code{\link[dymiumCore:Generic]{dymiumCore::Generic}} -> \code{\link[dymiumCore:Entity]{dymiumCore::Entity}} -> \code{\link[dymiumCore:Asset]{dymiumCore::Asset}} -> \code{Building}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_template}{\code{Building$data_template()}}
\item \href{#method-is_occupied}{\code{Building$is_occupied()}}
\item \href{#method-is_vacant}{\code{Building$is_vacant()}}
\item \href{#method-clone}{\code{Building$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="class">}\href{../../dymiumCore/html/Generic.html#method-class}{\code{dymiumCore::Generic$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="debug">}\href{../../dymiumCore/html/Generic.html#method-debug}{\code{dymiumCore::Generic$debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="is_dymium_class">}\href{../../dymiumCore/html/Generic.html#method-is_dymium_class}{\code{dymiumCore::Generic$is_dymium_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="message">}\href{../../dymiumCore/html/Generic.html#method-message}{\code{dymiumCore::Generic$message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_debug">}\href{../../dymiumCore/html/Generic.html#method-msg_debug}{\code{dymiumCore::Generic$msg_debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_info">}\href{../../dymiumCore/html/Generic.html#method-msg_info}{\code{dymiumCore::Generic$msg_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_trace">}\href{../../dymiumCore/html/Generic.html#method-msg_trace}{\code{dymiumCore::Generic$msg_trace()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_warn">}\href{../../dymiumCore/html/Generic.html#method-msg_warn}{\code{dymiumCore::Generic$msg_warn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="add_data">}\href{../../dymiumCore/html/Entity.html#method-add_data}{\code{dymiumCore::Entity$add_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="data">}\href{../../dymiumCore/html/Entity.html#method-data}{\code{dymiumCore::Entity$data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="generate_new_ids">}\href{../../dymiumCore/html/Entity.html#method-generate_new_ids}{\code{dymiumCore::Entity$generate_new_ids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_attr">}\href{../../dymiumCore/html/Entity.html#method-get_attr}{\code{dymiumCore::Entity$get_attr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_data">}\href{../../dymiumCore/html/Entity.html#method-get_data}{\code{dymiumCore::Entity$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_data_names">}\href{../../dymiumCore/html/Entity.html#method-get_data_names}{\code{dymiumCore::Entity$get_data_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_id_col">}\href{../../dymiumCore/html/Entity.html#method-get_id_col}{\code{dymiumCore::Entity$get_id_col()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_ids">}\href{../../dymiumCore/html/Entity.html#method-get_ids}{\code{dymiumCore::Entity$get_ids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_idx">}\href{../../dymiumCore/html/Entity.html#method-get_idx}{\code{dymiumCore::Entity$get_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_last_id">}\href{../../dymiumCore/html/Entity.html#method-get_last_id}{\code{dymiumCore::Entity$get_last_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_new_ids">}\href{../../dymiumCore/html/Entity.html#method-get_new_ids}{\code{dymiumCore::Entity$get_new_ids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="get_removed_data">}\href{../../dymiumCore/html/Entity.html#method-get_removed_data}{\code{dymiumCore::Entity$get_removed_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="has_attr">}\href{../../dymiumCore/html/Entity.html#method-has_attr}{\code{dymiumCore::Entity$has_attr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="ids_exist">}\href{../../dymiumCore/html/Entity.html#method-ids_exist}{\code{dymiumCore::Entity$ids_exist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="idx_exist">}\href{../../dymiumCore/html/Entity.html#method-idx_exist}{\code{dymiumCore::Entity$idx_exist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="n">}\href{../../dymiumCore/html/Entity.html#method-n}{\code{dymiumCore::Entity$n()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="print">}\href{../../dymiumCore/html/Entity.html#method-print}{\code{dymiumCore::Entity$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="print_data">}\href{../../dymiumCore/html/Entity.html#method-print_data}{\code{dymiumCore::Entity$print_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="remove">}\href{../../dymiumCore/html/Entity.html#method-remove}{\code{dymiumCore::Entity$remove()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Entity" data-id="summary">}\href{../../dymiumCore/html/Entity.html#method-summary}{\code{dymiumCore::Entity$summary()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Asset" data-id="get_owner">}\href{../../dymiumCore/html/Asset.html#method-get_owner}{\code{dymiumCore::Asset$get_owner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Asset" data-id="get_owner_id_col">}\href{../../dymiumCore/html/Asset.html#method-get_owner_id_col}{\code{dymiumCore::Asset$get_owner_id_col()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Asset" data-id="initialise_data">}\href{../../dymiumCore/html/Asset.html#method-initialise_data}{\code{dymiumCore::Asset$initialise_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Asset" data-id="initialize">}\href{../../dymiumCore/html/Asset.html#method-initialize}{\code{dymiumCore::Asset$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Asset" data-id="set_owner">}\href{../../dymiumCore/html/Asset.html#method-set_owner}{\code{dymiumCore::Asset$set_owner()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_template"></a>}}
\subsection{Method \code{data_template()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Building$data_template()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_occupied"></a>}}
\subsection{Method \code{is_occupied()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Building$is_occupied(ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_vacant"></a>}}
\subsection{Method \code{is_vacant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Building$is_vacant(ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Building$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
