% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingIndustrial.R
\name{BuildingIndustrial}
\alias{BuildingIndustrial}
\title{Industrial building class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Create dwelling class, extended Agent class.
}
\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{is_occupied(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is occupied
\item \code{is_vacant(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is vacant.
}
}

