% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildingResidential.R
\name{BuildingResidential}
\alias{BuildingResidential}
\title{Building Residential class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Building}<-\link{Asset}<-\link{Entity}.}
\description{
For storing methods and fields related to residential buildings.
}
\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{is_occupied(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is occupied
\item \code{is_vacant(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is vacant.
}
}

