% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{calculate_gene_signature_scores}
\alias{calculate_gene_signature_scores}
\title{Calculates scores for gene signature}
\usage{
calculate_gene_signature_scores(exp_matrix, dif_exp_results, signature_genes)
}
\value{
a list, which elements are named as signature name and
each element itself correspond to a named vector, where values are
scores (log2FC weighted expression) of a given signature in samples (TCGA case IDs) and names are TCGA case IDs
}
\description{
Calculate gene signature score as an average weighed expression for
selected genes, where weights are corresponding gene log2FC
}
