% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{calculate_roc_per_pw}
\alias{calculate_roc_per_pw}
\title{Calculate AUC ROC per each pathway based on pathways activation levels (PAL) and return ROC curve chart}
\usage{
calculate_roc_per_pw(
  sample_names,
  is_mutated,
  pathways_names,
  pw_activation_score,
  gene,
  aa_change = NULL
)
}
\arguments{
\item{sample_names}{a character vector, containing TCGA patient case id codes}

\item{is_mutated}{a character vector, containing either "Control" or "Mutated" values}

\item{pathways_names}{a character vector, contatining pathways names per each \code{sample_name}}

\item{pw_activation_score}{a numeric vector, containing PAL scores per each \code{sample_name} and \code{pathways_names} pair}

\item{gene}{SYMBOL of analysed gene}
}
\value{
list of arranged ROC plots for each of pathways
}
\description{
Calculate AUC ROC per each pathway based on pathways activation levels (PAL) and return ROC curve chart
}
