% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_entity_ids}
\alias{check_entity_ids}
\alias{assert_entity_ids}
\alias{test_entity_ids}
\alias{expect_entity_ids}
\title{Check if ids exist in Entity.}
\usage{
check_entity_ids(x, ids, include_removed_data = FALSE, informative = TRUE)

assert_entity_ids(
  x,
  ids,
  include_removed_data = FALSE,
  informative = TRUE,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_entity_ids(x, ids, include_removed_data = FALSE, informative = TRUE)

expect_entity_ids(
  x,
  ids,
  include_removed_data = FALSE,
  informative = TRUE,
  info = NULL,
  label = vname(x)
)
}
\arguments{
\item{x}{\link{Entity}\cr
an R6 \link{Entity} object}

\item{ids}{(\code{logical(1)})\cr
Ids of Entity to check.}

\item{include_removed_data}{(\code{any})\cr
Should the removed data of \code{x} be checked.}

\item{informative}{(\code{logical(1)})\cr
Whether to return the missing ids in error message.}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{add}{\link[checkmate:AssertCollection]{checkmate::AssertCollection}\cr
Collection to store assertions. See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}

\item{info}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{label}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}
}
\value{
(\code{logical(1)}).
}
\description{
Check if ids exist in Entity.
}
