% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{check_module}
\alias{check_module}
\title{Check if a module exists in a remote repository}
\usage{
check_module(name, repo = dymiumModulesRepo)
}
\arguments{
\item{name}{name of the module to check.}

\item{repo}{A GitHub repository to look for modules. By default, this uses 'dymium-org/dymiumModules'.}
}
\value{
a logical value.
}
\description{
Check if a module exists in a remote repository
}
\examples{

\dontrun{
  check_module('test')
}
}
