% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{check_module_version}
\alias{check_module_version}
\title{Check the existence of a module version.}
\usage{
check_module_version(name, repo = dymiumModulesRepo, version)
}
\arguments{
\item{name}{name of the module.}

\item{repo}{A GitHub repository to look for modules. By default, this uses 'dymium-org/dymiumModules'.}

\item{version}{a character. For example, if you would like to check
for version 0.1.0 type it as a character '0.1.0'.}
}
\value{
a logical value
}
\description{
Check the existence of a module version.
}
\examples{

\dontrun{
  check_module_version('test', version = '0.0.1')
}
}
