% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-event.R
\name{check_required_models}
\alias{check_required_models}
\title{Check for the required models in the model argument.}
\usage{
check_required_models(x, names)
}
\arguments{
\item{x}{a model object.}

\item{names}{names of the required models}
}
\value{
TRUE if all the required models exist else throws an error.
}
\description{
This function is meant to be used inside an Event function. It checks if the
model argument has all the required models.
}
