% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-helpers.R
\name{check_supported_model}
\alias{check_supported_model}
\title{Check if a model is supported by Transition}
\usage{
check_supported_model(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
TRUE if it is supported otherwise an error will be thrown.
}
\description{
Check if a model is supported by Transition
}
\examples{

check_supported_model(list(1))
}
