% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_transition_supported_model}
\alias{check_transition_supported_model}
\alias{assert_transition_supported_model}
\alias{test_transition_supported_model}
\alias{expect_transition_supported_model}
\title{Check if an argument is an object that is supported by \link{Transition}}
\usage{
check_transition_supported_model(x)

assert_transition_supported_model(
  x,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_transition_supported_model(x)

expect_transition_supported_model(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value if TRUE, if false then returns an error message
}
\description{
Check if an argument is an object that is supported by \link{Transition}
}
