% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clear}
\alias{clear}
\title{Deep cleaning of workspace}
\usage{
clear()
}
\value{
Output from \code{gc()}
}
\description{
Removes all variables in workspace and forces garbage collection.
}
\details{
Uses \code{rm()} to remove all variables from \code{ls()} and collects garbage with \code{gc()}
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{rm}} \code{\link{ls}} \code{\link{gc}}
}

