% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{code_packages}
\alias{code_packages}
\title{Development Packages}
\format{Character vector}
\usage{
code_packages
}
\description{
Names of biconductor packages, which should be loaded when developing.
}
\details{
The list of packages includes:

Development: Rccp, devtools, roxygen2, pryr, profr

RStudio: rstudioapi, manipulate

Terminal: setwidth, colorout
}
\author{
Malte Thodberg
}
\keyword{datasets}

