% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{colSds}
\alias{colSds}
\title{Column-wise standard deviations}
\usage{
colSds(d)
}
\arguments{
\item{d}{dataframe or matrix}
}
\value{
numeric vector
}
\description{
Calculates standard deviation for every column in dataframe or matrix
}
\details{
Wraps sd() with apply()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{apply}} \code{\link{sd}}
}

