% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{colVars}
\alias{colVars}
\title{Column-wise variances}
\usage{
colVars(d)
}
\arguments{
\item{d}{dataframe or matrix}
}
\value{
numeric vector
}
\description{
Calculates variance for every column in dataframe or matrix
}
\details{
Wraps var() with apply()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{apply}} \code{\link{var}}
}

