% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity-functions.R
\name{combine_histories}
\alias{combine_histories}
\title{Combine history data of Entities into a single data.frame.}
\usage{
combine_histories(x)
}
\arguments{
\item{x}{a R6 object that inherits \link{Container} such as \link{World}.}
}
\value{
a data.table with five columns: time, created_timestamp, event, id, entity
}
\description{
Combine history data of Entities into a single data.frame.
}
\examples{

create_toy_world()
Ind <- world$get("Individual")
add_history(Ind, ids = sample(Ind$get_ids(), 10), event = "event1", time = 1)
combine_histories(world)
}
