% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{condense_rows}
\alias{condense_rows}
\title{Condense rows}
\usage{
condense_rows(id, target, names = NULL)
}
\arguments{
\item{id}{(\code{integer()}) a integer vector contains ids.}

\item{target}{(\code{integer()}) a integer vector contains values with the same length as \code{id}.}

\item{names}{(Default as \code{NULL})}
}
\value{
a data.table with two columns, \code{id} and \code{target}.
}
\description{
Condense two vectors - where the first vector is an id vector and the second
vector is an target (which must be unique) -  into a data.table with two
columns (id and target).
}
