% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContainerGeneric.R
\name{ContainerGeneric}
\alias{ContainerGeneric}
\title{ContainerGeneric}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[dymiumCore:Generic]{dymiumCore::Generic}}
\description{
ContainerGeneric is the base class for \link{Container}.
}
\section{Construction}{
\preformatted{x <- ContainerGeneric$new()
}
\itemize{
\item NULL\cr
}
}

\section{Public Fields}{

\itemize{
\item \code{Cont}\cr
A named list that contains objects of any type.
}
}

\section{Public Methods}{

\itemize{
\item \code{add(x)}\cr
(any objects)\cr
Add \code{x} to self$Cont.
\item \code{check_pos(x)}\cr
(\code{integer(1)}) -> (\code{logical(1)})\cr
Check if the position in \code{x} exists in \code{self$Cont}.
\item \code{get(x)}\cr
(\code{integer(1)}) -> (an object)\cr
Get a object in \code{x} element of \code{self$Cont}.
\item \code{n()}\cr
() -> (\code{integer(1)})\cr
Returns the length of \code{self$Cont}.
\item \code{print()}\cr
\item \code{remove(x)}\cr
(\code{integer(1)})\cr
Remove the object at the \code{x} element of \code{self$Cont}.
\item \code{unpack(target)}\cr
(\link[dymiumCore:Container]{dymiumCore::Container})\cr
Add all the objects in \code{self$Cont} to another \code{target} Container.
\item \code{summary()}\cr
Abstract method.
}
}

\section{Super class}{
\code{\link[dymiumCore:Generic]{dymiumCore::Generic}} -> \code{ContainerGeneric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContainerGeneric$new()}}
\item \href{#method-print}{\code{ContainerGeneric$print()}}
\item \href{#method-summary}{\code{ContainerGeneric$summary()}}
\item \href{#method-add}{\code{ContainerGeneric$add()}}
\item \href{#method-get}{\code{ContainerGeneric$get()}}
\item \href{#method-unpack}{\code{ContainerGeneric$unpack()}}
\item \href{#method-n}{\code{ContainerGeneric$n()}}
\item \href{#method-remove}{\code{ContainerGeneric$remove()}}
\item \href{#method-check_pos}{\code{ContainerGeneric$check_pos()}}
\item \href{#method-clone}{\code{ContainerGeneric$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="class">}\href{../../dymiumCore/html/Generic.html#method-class}{\code{dymiumCore::Generic$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="debug">}\href{../../dymiumCore/html/Generic.html#method-debug}{\code{dymiumCore::Generic$debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="is_dymium_class">}\href{../../dymiumCore/html/Generic.html#method-is_dymium_class}{\code{dymiumCore::Generic$is_dymium_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="message">}\href{../../dymiumCore/html/Generic.html#method-message}{\code{dymiumCore::Generic$message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_debug">}\href{../../dymiumCore/html/Generic.html#method-msg_debug}{\code{dymiumCore::Generic$msg_debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_info">}\href{../../dymiumCore/html/Generic.html#method-msg_info}{\code{dymiumCore::Generic$msg_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_trace">}\href{../../dymiumCore/html/Generic.html#method-msg_trace}{\code{dymiumCore::Generic$msg_trace()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_warn">}\href{../../dymiumCore/html/Generic.html#method-msg_warn}{\code{dymiumCore::Generic$msg_warn()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$add(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$get(pos)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unpack"></a>}}
\subsection{Method \code{unpack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$unpack(target)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n"></a>}}
\subsection{Method \code{n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$n()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$remove(pos)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_pos"></a>}}
\subsection{Method \code{check_pos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$check_pos(pos)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContainerGeneric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
