% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convert_ensembl_to_symbol_gene_names}
\alias{convert_ensembl_to_symbol_gene_names}
\title{Maps ENSEMBL gene names into HUGO (HGNC / SYMBOL) format and vice versa}
\usage{
convert_ensembl_to_symbol_gene_names(
  gene_names,
  gene_format = c("ensembl_gene_id", "hgnc_symbol"),
  counts,
  hsmart_ann,
  ...
)
}
\arguments{
\item{gene_names}{a vector, containing gene names in ENSEMBL format}

\item{gene_format}{a str, specifying either \code{ensembl_gene_id} or \code{hgnc_symbol} input gene format}

\item{counts}{a vector, containig raw counts for gene expression}

\item{hsmart_ann}{a R object, containig HSMART ENSEMBL genes annotation}
}
\value{
a data frame with columns: \code{ensembl_gene_id}, \code{SYMBOL}, \code{counts}
}
\description{
Maps ENSEMBL gene names into HUGO (HGNC / SYMBOL) format and vice versa
}
