% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{core_packages}
\alias{core_packages}
\title{Common Packages}
\usage{
core_packages()
}
\value{
Character vector
}
\description{
Simply return a vector with names of most common packages. Let's see if this is pushed
}
\details{
The list of packages includes:

Data manipulation: magrittr, readr, tidyr, dplyr

Special data formats: stringr, lubridate

Plotting: grid, gridExtra, ggplot2, GGally, ggthemes, ggExtra, RColorBrewer, VennDiagram, pheatmap, wesanderson

Performance: parallel
}
\author{
Malte Thodberg
}

