% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{cRamp}
\alias{cRamp}
\title{Manual color gradient from values.}
\usage{
cRamp(x, palette = "rainbow", levels = 10)
}
\arguments{
\item{x}{numeric vector}

\item{palette}{color palette (rainbow, terrain, heat or topo)}

\item{levels}{number of different colors.}
}
\value{
character vector containing RGB values
}
\description{
Given a numeric vector, this function creates a vector of colors scaled to the values.
}
\details{
This function is usefull for creating color gradients for use with base plotting.
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{rainbow}} \code{\link{terrain.colors}} \code{\link{heat.colors}} \code{\link{topo.colors}}
}

