\name{create.output}
\alias{create.output}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday private }
\description{
  Takes the DATA object returned by the applicable function and creates the
  output parseable by the java side of the plugin.
}
\usage{
create.output(result)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{result}{ The DATA structure returned by an applicable function }
}
\details{
  Do not use this function directly within your R-code.
}
\value{
  none
}
\author{ Matthias Zschunke}
\keyword{ programming }% at least one, from doc/KEYWORDS