% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choice.R
\name{create_choice_table}
\alias{create_choice_table}
\title{prepend dymium.choice_table}
\usage{
create_choice_table(x)
}
\arguments{
\item{x}{any object.}
}
\value{
\code{x}
}
\description{
prepend dymium.choice_table
}
