% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{create_new_event}
\alias{create_new_event}
\title{Create a event script from an recommended event template}
\usage{
create_new_event(
  event_name,
  module_name,
  module_path,
  filename = NULL,
  author = NULL,
  author.email = NULL,
  module.version = "0.0.1",
  dymiumCore.version = utils::packageVersion("dymiumCore"),
  .overwrite = FALSE
)
}
\arguments{
\item{event_name}{\code{character(1)}\cr
Name of the event to be created. This will also be used as its event script name
(i.e. for example if 'age' was given then the script will be named as 'age.R')
and also as the function name: 'event_(module_name)_(event_name)'.}

\item{module_name}{\code{character(1)}\cr
Name of the module, this function will be named in this format:
'event_(module_name)_(event_name)'.}

\item{module_path}{\code{character(1)}\cr
Directory where the event script will be created.}

\item{filename}{\code{character(1)} or \code{NULL}\cr
The name of the script to be created. If \code{NULL} event_name is used as the filename.}

\item{author}{name of aithor(s)}

\item{author.email}{author's email}

\item{module.version}{module version eg: (0.0.x).}

\item{dymiumCore.version}{version of dymiumCore that was used to create the event.}

\item{.overwrite}{\code{logical(1)} default as FALSE.}
}
\value{
Invinsibly returns the path to the generated script.
}
\description{
Create a event script from an recommended event template
}
\examples{

create_new_event(event_name = "age",
                 module_name = "demography",
                 module_path = tempdir())
}
