% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{create_new_module}
\alias{create_new_module}
\title{Create new module}
\usage{
create_new_module(name, event, path)
}
\arguments{
\item{name}{module}

\item{event}{events}

\item{path}{path to create module}
}
\description{
Create new module
}
\examples{

create_new_module("populationDynamic", event = c("age", "birth", "death"), path = tempdir())
}
