% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-world.R
\name{create_toy_population}
\alias{create_toy_population}
\title{Create a loaded population object}
\usage{
create_toy_population()
}
\value{
assign \code{pop} to global env.
}
\description{
Create a loaded population object and assign to the global environment as \code{pop}.
}
