% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-world.R
\name{create_toy_world}
\alias{create_toy_world}
\title{create toy world}
\usage{
create_toy_world(add_toy_zones = TRUE)
}
\arguments{
\item{add_zone}{add \code{toy_zones} to world (default TRUE).}
}
\description{
Create a toy world (assigned to the global environment as \code{world}) for running tests
}
