% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackend.R
\name{DataBackend}
\alias{DataBackend}
\title{DataBackend class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
The base class of all DataBackends. \link{DataBackend} isn't supposed to be use directly.
It is meant to be implemented by the user. It is recommended that \link{DataBackendDataTable}
is to be used when adding a non-spatial data and \link{DataBackendSpatialFeature} for
spatial data.
}
\section{Construction}{
\preformatted{x <- DataBackend$new(.data)
}
\itemize{
\item \code{.data}\cr
A data object of any types.
}
}

\section{Fields}{

\itemize{
\item NULL\cr
}
}

\section{Public Methods}{

\itemize{
\item \code{add(x)}\cr
(\strong{Abstract})
\item \code{get(x)}\cr
(\strong{Abstract}) Return the data.
\item \code{view(x)}\cr
(\strong{Abstract}) Return the data interactively.
\item \code{remove(x)}\cr
(\strong{Abstract}) Remove the data.
\item \code{class(x)}\cr
Return the class of the data.
\item \code{n(x)}\cr
(\strong{Abstract}) Return the number of rows or entries in the data.
\item \code{get_removed()}\cr
Return the removed data.
}
}

\section{Private Fields}{

\itemize{
\item \code{.data}::\code{NULL}\cr
\item \code{.removed_data}::\code{NULL}\cr
}
}

