\name{devSVGMulti}
\alias{devSVGMulti}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{R for Mayday private}
\description{
   Wrapper function for devSVG to get this automatically
   loaded by the high level plots with the parameter 
   onefile=FALSE, indicating that there are more than one
   plots in the resulting svg-file.
}
\usage{
devSVGMulti()
}
%- maybe also 'usage' for other objects documented here.
\details{
	Do not use this function directly from your 
	\emph{applicable functions}.  
}
\value{
	none
}
%\references{  }
\author{ Matthias Zschunke}

\seealso{ \code{\link{devSVG}} in package RSvgDevice}
\keyword{ programming }% at least one, from doc/KEYWORDS
