% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{dif_exp_analysis}
\alias{dif_exp_analysis}
\title{Perform Differential Expression Analysis}
\usage{
dif_exp_analysis(exp_matrix, sample_ann, shring_results = FALSE)
}
\arguments{
\item{sample_ann}{a named vector, where names are TCGA sample names (case IDs)}

\item{exp_data}{a data frame, where rows are HGNC genes (HUGO)
and cols are TCGA sample names (case IDs)}
}
\value{
a DESeq2 results object
}
\description{
Perform Differential Expression Analysis
}
