% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-helpers.R
\name{dm_get_model}
\alias{dm_get_model}
\title{Helper - get model objects from Container}
\usage{
dm_get_model(x, model_names)
}
\arguments{
\item{x}{a \link{Container} object with \code{get_model()} method.}

\item{model_names}{a character vector}
}
\value{
a named \code{list} containing \link{Model} objects
}
\description{
Helper - get model objects from Container
}
\examples{

create_toy_world()
testModel <- list(a = 1)
world$add(testModel, "testModel")
dm_get_model(world, "testModel")
}
