% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-helpers.R
\name{dm_ref}
\alias{dm_ref}
\title{Get an object from a Container.}
\usage{
dm_ref(x, name)
}
\arguments{
\item{x}{a \link{Container} object.}

\item{name}{a \code{character} or a \link[R6:R6Class]{R6::R6Class} generator.}
}
\value{
a \link[R6:R6Class]{R6::R6Class} object.
}
\description{
Get an object from a Container.
}
\examples{

create_toy_world()
# both lines return the same object.
dm_ref(world, Individual)
dm_ref(world, "Individual")
}
