% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipeline.R
\name{dm_save}
\alias{dm_save}
\title{Save an instance of Entity or Container.}
\usage{
dm_save(x, dir = getOption("dymium.output_dir"))
}
\arguments{
\item{x}{an instance of \link{Entity} of \link{Container}.}

\item{dir}{By default this is set to the active scenario directory which can be
viewed by calling \code{getOption("dymium.output_dir")}. \code{x} will be saved as a .RDS
file with the name being the class name of \code{x}.}
}
\description{
This is part of the dm_* family of functions to be used inside a microsimulation
pipeline. This saves an instance of Entity or Container to a directory.
}
\examples{

if (FALSE) {
  # create a World object and assign it as 'world' to the global environment.
  create_toy_world()

  # this saves the 'world' object to the outputs folder of the active scenario folder.
  dm_save(world)
}
}
