% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{.get_sim_time}
\alias{.get_sim_time}
\title{get_sim_time}
\usage{
.get_sim_time()
}
\value{
a numerical value
}
\description{
get the simulation time from .DMevn$sim_time. This object is readable by all
objects in the package but only modifiable by instances of World-class through
\verb{$set_time(x)} method.
}
\examples{

.get_sim_time()
}
