% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_group_and_sort}
\alias{dt_group_and_sort}
\title{Group a column into a list column and sort by the group-by column}
\usage{
dt_group_and_sort(x, groupby_col, group_col, sort_order)
}
\arguments{
\item{x}{:: \code{\link[data.table:data.table]{data.table::data.table()}}}

\item{groupby_col}{:: \code{character(1)}
the column to use in 'by' in the data.table \code{x}.}

\item{group_col}{:: \code{character(1)}
the column to be grouped into a list column.}

\item{sort_order}{:: \code{vector()}
A vector of ordered values to sort the \code{groupby_col}. If a value in the
vector is missing from \code{groupby_col} in \code{x} its \code{group_col} will be \code{NA}.}
}
\value{
\code{\link[data.table:data.table]{data.table::data.table()}} with 2 columns: \code{sort_col}, \code{group_col}.
}
\description{
Group a column into a list column and sort by the group-by column
}
\note{
The returned \code{\link[data.table:data.table]{data.table::data.table()}} should have the same number of rows
as the length of \code{sort_col}.
In short, the if-statement makes sure that all emptied \code{group_col} will be \code{NA} not \code{NULL}

For cases where some \code{sort_order} are not in \code{groupby_col} of \code{x}
those elements in \code{sort_order} will have their values in \code{groupby_col} as \code{NULL}
}
