% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{element_wise_expand_lists}
\alias{element_wise_expand_lists}
\title{element_wise_expand_lists}
\usage{
element_wise_expand_lists(l1, l2)
}
\arguments{
\item{l1}{first list}

\item{l2}{second list}
}
\value{
data.frame with two columns, Var1 and Var2.
}
\description{
expand two lists into a dataframe. l1 and l2 can be a nested list.
}
\examples{
 l1 <- list(1,2,3,4)
 l2 <- list(1,2,3,4)
 element_wise_expand_lists(l1,l2)

 l1 <- list(1,2,3,4)
 l2 <- list(1,2,3,c(1,3))
 element_wise_expand_lists(l1,l2)
}
