% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{export_analysis_plot}
\alias{export_analysis_plot}
\title{Save plot of provided analysis step}
\usage{
export_analysis_plot(
  filename,
  plot,
  path,
  scale,
  width,
  height,
  units,
  load_fonts = FALSE
)
}
\description{
Export ggplot2 object to desired location in vector (\code{.pdf}) and raster (\code{.tiff}) formats
}
