\name{extract.colors}
\alias{extract.colors}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday user functions  }
\description{
  Extract the colors for a list of probes from the DATA\verb+!+probelists object.
}
\usage{
extract.colors(probes, PL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{probes}{ A vector of probe ids.}
  \item{PL}{ The probelists substructure of DATA. }
}
\details{
  Use this in your applicable function.
}
\value{
 A vector of the corresponding colors, each belonging to the given
 probe id.
}

\author{ Matthias Zschunke }

\keyword{ programming }% at least one, from doc/KEYWORDS
