\name{extract.probelist.names}
\alias{extract.probelist.names}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday user function }
\description{
  Extract the names of all probelists given in the DATA object.
}
\usage{
extract.probelist.names(DATA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DATA}{ The DATA object. }
}
\details{
  You can use this funcition from your \emph{applicable functions}.
}
\value{
	A vector of strings containing all Probelist identifier.
}
%\references{}
\author{ Matthias Zschunke }
%\note{ ~~further notes~~ }

\keyword{ programming }% at least one, from doc/KEYWORDS
