% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipeline.R
\name{extract_data}
\alias{extract_data}
\alias{extract_data.World}
\alias{extract_data.Entity}
\alias{extract_data.DataBackendDataFrame}
\title{Extract data from Entity objects}
\usage{
extract_data(x)

\method{extract_data}{World}(x)

\method{extract_data}{Entity}(x)

\method{extract_data}{DataBackendDataFrame}(x)
}
\arguments{
\item{x}{a \link{World} or \link{Entity} object.}
}
\value{
a named list of data from their DataBackend objects.
}
\description{
Extract data from Entity objects
}
\examples{

create_toy_world()
extract_data(world)
extract_data(world$entities$Individual)
}
