% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filter_non_common_samples}
\alias{filter_non_common_samples}
\title{Filter non-common samples}
\usage{
filter_non_common_samples(exp_matrix, sample_ann)
}
\arguments{
\item{sample_ann}{a named vector, where names are TCGA sample names (case IDs)}

\item{exp_data}{a data frame, where rows are HGNC genes (HUGO)
and cols are TCGA sample names (case IDs)}
}
\description{
Update in a global environment \code{exp_matrix} and \code{sample_ann}
by versions having matching sample names
}
