% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_population}
\alias{generate_population}
\title{Randomly generate households and individuals}
\usage{
generate_population(n_households = 10, max_hhsize = 6)
}
\arguments{
\item{n_households}{(\code{integer(1)}) number of households to be generated.}

\item{max_hhsize}{(\code{integer(1)}) the maximum size of households that can be generated.}
}
\value{
a list of data.tables
}
\description{
This function returns a list of randomly generated housheold and individual
data.tables which comprise of these columns for individuals: pid, hid, age, sex.
And these following columns for households: hid, hhsize. Note that, the total
number of individuals cannot be garantee to be the same in every run.
}
\examples{

synth_pop <- generate_population()
}
