% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{Generic}
\alias{Generic}
\title{Generic class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Can be used as a template for other objects. This contains useful utility
functions for abstraction objects. See the public members (methods).
}
\section{Construction}{
\preformatted{Obj <- Generic$new(...)
}
}

\section{Fields}{

\itemize{
\item \code{NULL}
}
}

\section{Methods}{

\itemize{
\item \code{debug()}\cr
() -> \code{self & private}\cr
Makes debugging the object easier by assigning self and private to the
global environment of the current seesion.
\item \code{message(x, ...)}\cr
Print \code{message} onto the console with the object classname append in front of
the message. To turn this off, set options(dymium.agent.verbose = FALSE).
}
}

