% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{get_all_module_files}
\alias{get_all_module_files}
\title{Get all files from all modules in a repository.}
\usage{
get_all_module_files(repo = dymiumModulesRepo)
}
\arguments{
\item{repo}{A GitHub repository to look for modules. By default, this uses 'dymium-org/dymiumModules'.}
}
\value{
a character vector.
}
\description{
Get all files from all modules in a repository.
}
