% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_drug_scores_sample_annotations}
\alias{get_drug_scores_sample_annotations}
\title{Preprocess Drug Scores database into sample annotation}
\usage{
get_drug_scores_sample_annotations(
  drug_scores_db_path = "data/drug_scores/ds_balanced.csv",
  drug_name = "Cetuximab",
  return_ranks = FALSE,
  sample_list_path = NULL
)
}
\arguments{
\item{sample_list_path}{optional. Path to \code{.txt} file containing a list of TCGA patient codes to subset from the expression matrix.}
}
\description{
Preprocess Drug Scores database into sample annotation
}
