% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_gene_signature_sample_annotations}
\alias{get_gene_signature_sample_annotations}
\title{Preprocess DESeq2 results into sample annotation for gene signature of significantly differentially expressed genes}
\usage{
get_gene_signature_sample_annotations(
  dif_exp_results_path,
  candidate_genes,
  expression_matrix,
  pCutoff = 0.05,
  FCcutoff = 0.5,
  n_genes_per_signature = NULL
)
}
\arguments{
\item{dif_exp_results_path}{path to .RDS object corresponding to the DESeq2 results data frame}

\item{candidate_genes}{character vector, containing unique HUGO (HGNC) name for genes,
that will be used to form a signature if their differential expression passes \code{pCutoff} and \code{pCutoff} thresholds}

\item{expression_matrix}{a data frame containing preprocessed values for gene expression. See \code{load_exp_matrix} function}

\item{n_genes_per_signature}{numeric. If specified it will generate multiple signatures from a set of passed \code{candidate_genes},
where each signature will have a size of specified number.}
}
\value{
a list, which elements are named as signature name and
each element itself correspond to a named vector, where values are
scores (log2FC weighted expression) of a given signature in samples (TCGA case IDs) and names are TCGA case IDs
}
\description{
Preprocess DESeq2 results into sample annotation for gene signature of significantly differentially expressed genes
}
