% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity-functions.R
\name{get_history}
\alias{get_history}
\title{Get Entity history}
\usage{
get_history(x)
}
\arguments{
\item{x}{An \link{Entity} or \link{Container} object}
}
\value{
a list of data.table or \code{NULL}.
}
\description{
Get the history data of a given Entity. If a \link{Container} object is given then
this will return a list of data.tables or NULL if an \link{Entity} in the \link{Container}
doesn't have any history data.
}
\examples{

create_toy_world()

add_history(world$get("Individual"),
            ids = 1:100,
            event = "test_event1",
            time = 1)

get_history(world)
get_history(world$get("Individual"))
}
