% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{get_log}
\alias{get_log}
\alias{get_log.Generic}
\alias{get_log.Container}
\title{Get log from Generic}
\usage{
get_log(x)

\method{get_log}{Generic}(x)

\method{get_log}{Container}(x)
}
\arguments{
\item{x}{a \link{Generic} object.}
}
\value{
a log data.table.
}
\description{
Get log from Generic
}
