% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{get_module_versions}
\alias{get_module_versions}
\title{Get all version numbers of a module}
\usage{
get_module_versions(name, repo = dymiumModulesRepo)
}
\arguments{
\item{name}{name of the module.}

\item{repo}{A GitHub repository to look for modules. By default, this uses 'dymium-org/dymiumModules'.}
}
\value{
a character vector.
}
\description{
Get all version numbers of a module
}
\examples{
\dontrun{
get_module_versions("demography")
}
}
