% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{get_modules}
\alias{get_modules}
\title{Get the names of available modules from a remote repository}
\usage{
get_modules(repo = dymiumModulesRepo)
}
\arguments{
\item{repo}{A GitHub repository to look for modules. By default, this uses 'dymium-org/dymiumModules'.}
}
\value{
a character vector.
}
\description{
Get the names of available modules from a remote repository
}
\examples{
\dontrun{
get_modules()
}
}
