% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_pathways_activation_sample_annotations}
\alias{get_pathways_activation_sample_annotations}
\title{Preprocess Pathways Activation database into sample annotation}
\usage{
get_pathways_activation_sample_annotations(
  pathways_activation_db_path = "data/pathways_activation_scores/EGF_PWs.xlsx",
  return_dataframe = FALSE,
  return_ranks = FALSE
)
}
\value{
a list, which elements are named as pathway names and
each element itself correspond to a named vector, where values are
activations of a given pathway in samples (TCGA case IDs) and names are TCGA case IDs
}
\description{
Preprocess Pathways Activation database into sample annotation
}
