% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{get_sim_time}
\alias{get_sim_time}
\alias{.get_sim_time}
\title{Get the current timestep of the simulation

This function returns the time attribute of a \link{World} object, if any, otherwise
it returns 0L. Note that, this should not be set through \code{options()}, instead
it should be set by a \link{World} object through \code{World$set_time()}.}
\usage{
get_sim_time()

.get_sim_time()
}
\value{
a numerical value
}
\description{
Get the current timestep of the simulation

This function returns the time attribute of a \link{World} object, if any, otherwise
it returns 0L. Note that, this should not be set through \code{options()}, instead
it should be set by a \link{World} object through \code{World$set_time()}.
}
\examples{

get_sim_time()
}
