% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthr-package.R
\docType{package}
\name{healthr-package}
\alias{healthr}
\alias{healthr-package}
\title{healthr: Industrial Health Monitoring Framework}
\description{
A modular framework for real-time industrial process monitoring,
anomaly detection, and predictive analytics using time-series forecasting.
}
\author{
\strong{Maintainer}: Abraham Jaimes \email{abraham.jaimes.mx@gmail.com}

}
\keyword{internal}
