% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{hello1}
\alias{hello1}
\title{Test function}
\usage{
hello1(x, y)
}
\arguments{
\item{x}{the first number}

\item{y}{the second number}
}
\value{
list
}
\description{
Print hello world and return list
}
\details{
Comments for basic function
}
\author{
Malte Thodberg
}

