% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household-formation.R
\name{household_formation}
\alias{household_formation}
\title{household formation functions}
\arguments{
\item{Pop}{a \link{Population} object}

\item{mapping}{a data.table with two columns {ind_id, hh_id}. hh_id is
only required when type == "join" or "randomjoin".}
}
\value{
\code{NULL}
}
\description{
This family of functions change the state of individual data and
household data of the input Pop-class object.
}
