% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Household.R
\name{Household}
\alias{Household}
\title{Household class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Agent}<-\link{Entity}.}
\description{
Create Household class, extended Agent class.
}
\section{Construction}{
\preformatted{Household$new()
}
}

\section{Methods}{


Inherits all methods in \link{Agent}.
}

\section{Overwritten methods}{

\itemize{
\item \code{add_new_agents(data, n)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}, \code{integer(1)}) -> \code{NULL}\cr
Add new households using data or generate \code{n} emptied households. If \code{data}
is given \code{n} will be ignored.
\item \code{initialise_data(data, id_col)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}, \code{character(1)}) -> \code{NULL}\cr
Store data to data while making sure that \code{data} meets all the
requirements.
}
}

\concept{Agent}
