% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household-formation.R
\name{household_formation}
\alias{household_formation}
\title{household formation functions}
\usage{
household_formation(Pop, model = NULL, mapping, type)
}
\arguments{
\item{Pop}{\link{Population}.}

\item{model}{a named list that contains a distribution of household sizes that
the new individuals/households are going to join. For example, \verb{list(1 = 0.5, 2 = 0.5)}.}

\item{mapping}{a \code{\link[data.tabe:data.table]{data.tabe::data.table()}} that contains two columns: \code{ind_id}
and \code{hh_id}.}

\item{type}{one of 'join', 'randomjoin' or 'new'.}
}
\value{
\code{NULL}
}
\description{
This family of functions change the state of individual data and
household data of the input Pop-class object.
}
