% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{ht}
\alias{ht}
\title{Head and tail}
\usage{
ht(d, n = 5)
}
\arguments{
\item{d}{dataframe or matrix}

\item{n}{number of rows from head and tail}
}
\value{
dataframe a matrix with n*2 rows from head and tail
}
\description{
Show the first and last few rows of a dataframe or matrix.
}
\details{
Wraps head and tail with rbind for convenience.
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{head}} \code{\link{tail}}
}

