% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity-functions.R
\name{inspect}
\alias{inspect}
\title{Inspect Entity}
\usage{
inspect(entity, ids, related_entity = NULL, verbose = TRUE)
}
\arguments{
\item{entity}{an \link{Entity} object or its inheritance object}

\item{ids}{a integer vactor that contains the ids of the entities to be inspected}

\item{related_entity}{an \link{Entity} object or its inheritance object which must
be related to \code{entity} but not the same as \code{entity}. For example, \code{related_entity}
can be \link{Household} if \code{entity} is \link{Individual}.}

\item{verbose}{:: \code{logical(1)}
Should the data be printed to console.}
}
\value{
a named ist of data.tables.
}
\description{
Return and print to console the attribute data of entities within an \link{Entity}
object or its inheritance object. A related entity object can also be queried
at the same time.
}
\examples{

create_toy_world()

inspect(
  entity = world$get("Individual"),
  ids = c(1:3),
  related_entity = world$get("Household")
)
}
