% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-event.R
\name{is_scheduled}
\alias{is_scheduled}
\title{has the event been scheduled?}
\usage{
is_scheduled(time_steps)
}
\arguments{
\item{time_steps}{a numeric vector}
}
\value{
a logical value
}
\description{
has the event been scheduled?
}
