\name{jpeg100}
\alias{jpeg100}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{R for Mayday private}
\description{
   Wrapper function for the jpeg device that simply calls
   the jpeg device with quality=100 option.
}
\usage{
jpeg100()
}
%- maybe also 'usage' for other objects documented here.
\details{
	Do not use this function directly from your 
	\emph{applicable functions}.  
}
\value{
	none
}
%\references{  }
\author{ Matthias Zschunke}

\seealso{ \code{\link{jpeg}} in package grDevices}
\keyword{ programming }% at least one, from doc/KEYWORDS
