% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{keep}
\alias{keep}
\title{The Keep Function}
\usage{
keep(..., x = c())
}
\arguments{
\item{...}{Variables that should stay in the environment}

\item{x}{A character vector of variables which you wish to keep}
}
\description{
Modified from https://github.com/flor3652/myStuff/blob/master/R/keep.R. This function is designed to keep a given list of variables (as opposed to the rm function, which is designed to remove a list of variables). In addition to the original code, it also forces a gc() garbage collection.
}
\details{
The function uses the \code{ls} command to find the variables that are defined in the global environment. It then deletes all variables that are not given as an argument.
}
\author{
Michael Floren
}
\seealso{
\code{\link{rm}} \code{\link{ls}} \code{\link{gc}}
}

