% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{l}
\alias{l}
\title{Short length}
\usage{
l()
}
\value{
Similar to length()
}
\description{
Shorter call to length
}
\details{
Similar to length()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{length}}
}

