% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_exp_matrix}
\alias{load_exp_matrix}
\title{Load preprocessed expression matrix}
\usage{
load_exp_matrix(
  exp_matrix_path = "data/exp_matrix.txt",
  drop_dupl_genes = FALSE,
  sample_list_path = NULL,
  quantile_normalization = TRUE,
  log10_scalling = TRUE
)
}
\arguments{
\item{exp_matrix_path}{path to processed expression matrix obtained from \code{source_data_to_matrix}}

\item{drop_dupl_genes}{whether to remove duplicated gene names (ex. MIR*, LINC*). By default is \code{FALSE}}

\item{sample_list_path}{optional. Path to \code{.txt} file containing a list of TCGA patient codes to subset from the expression matrix.}

\item{quantile_normalization}{bool. Whether to conduct Quantile Normalization of expression data}

\item{log10_scalling}{bool. Wherher to convert counts to log10 values}
}
\value{
expression matrix where cols are sample names and rows are SYMBOL genes
}
\description{
Load preprocessed expression matrix
}
