% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_and_replace}
\alias{lookup_and_replace}
\title{lookup_and_replace}
\usage{
lookup_and_replace(data, lookup_table, cols, id_col = NULL)
}
\arguments{
\item{data}{a data.table/data.frame to be replaced.}

\item{lookup_table}{a data.table/data.frame contains two columns which must be
named as \code{.key} and \code{value}. The reason for requiring a dot infront of \code{key}
is that \code{key} is a reserved column name in \code{data.table} meaning no columns
can be named \code{key}. Also the \code{.key} column should have the same data type as
\code{cols_to_lookup} columns.}

\item{cols}{column names to be lookup and replace}

\item{id_col}{a column with unique value to be used in unnesting \code{data}. This
is not required if \code{cols_to_lookup} doesn't contain any list columns.}
}
\value{
data.table with new values
}
\description{
Replace all values in a data.frame using a lookup table. Maximum of one list
column can be include in \code{cols_to_lookup}.
}
