% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_and_replace2}
\alias{lookup_and_replace2}
\title{Look up and replace values in columns.}
\usage{
lookup_and_replace2(x, cols, mapping)
}
\arguments{
\item{x}{a data.frame}

\item{cols}{a character vector of column names in \code{x} to be replaced.}

\item{mapping}{a data.frame contains two columns which are \code{.key} and \code{.value}}
}
\value{
a data.frame
}
\description{
Replace keys with their values.
}
