% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\alias{makeModel.train}
\alias{makeModel.WrappedModel}
\title{make Model}
\usage{
makeModel(model, ...)

\method{makeModel}{train}(model, preprocessing_fn = NULL)

\method{makeModel}{WrappedModel}()
}
\arguments{
\item{model}{a model object. See \code{SupportedTransitionModels()}.}

\item{...}{dots.}
}
\value{
a \code{Model} object.
}
\description{
Make a light weight model object that can be used by \code{transition()}, \code{TransitionClassification}, \code{TransitionRegression}.
}
