% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingMarketOptimal.R
\name{MatchingMarketOptimal}
\alias{MatchingMarketOptimal}
\title{Optimal Matching Market}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{MatchingMarket}.}
\description{
For general detail See \link{MatchingMarket}. This class solves matching
problems using optimal matching approaches such as one-to-one matching, many-to-one
one-sided matching, two-sided matching, etc. This matching strategy assumes
agents have perfect information of the market.
}
\section{Fields}{

See \link{MatchingMarket}.
}

\section{Methods}{

See \link{MatchingMarket}.
}

\seealso{
Other MatchingMarket: 
\code{\link{MatchingMarketStochastic}}
}
\concept{MatchingMarket}
