% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchingMarketStochastic.R
\name{MatchingMarketStochastic}
\alias{MatchingMarketStochastic}
\title{Stochastic Matching Market}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{MatchingMarket}.}
\description{
For general detail See \link{MatchingMarket}. This class solves a matching
problem using a stochastic strategy such as randomly pick based on probabilistic scores
or ranking. This stretegy mimics cases where agents have imperfect imformation about
the market hence it only awares of a subset of all the available options in the market.
}
\section{Fields}{

See \link{MatchingMarket}.
}

\section{Methods}{

See \link{MatchingMarket}.
}

\seealso{
Other MatchingMarket: 
\code{\link{MatchingMarketOptimal}}
}
\concept{MatchingMarket}
