% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimulate.R
\name{microsimulate}
\alias{microsimulate}
\alias{microsimulate.train}
\alias{microsimulate.glm}
\title{Microsimulate a choice situation}
\usage{
microsimulate(model, ...)

\method{microsimulate}{train}(model, newdata, target)

\method{microsimulate}{glm}(model, newdata, target)
}
\arguments{
\item{model}{a \link{Model} object or an object in \code{\link[=SupportedTransitionModels]{SupportedTransitionModels()}}.}

\item{...}{dots}

\item{newdata}{a data.frame object to use for making prediction.}

\item{target}{a \link{Target} object or a named list.}
}
\value{

}
\description{
This function simulates a choice situation using a model object and data to
predict the probability.
}
\examples{

TODO
}
