\name{miogroups.read}
\alias{miogroups.read}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday private }
\description{
  Read the mio group ids.
}
\usage{
miogroups.read(Lines)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Lines}{ The lines from the input file corresponding to the
  miogroups section.}
}
\details{
	Do not use this function in your code directly!
}
\value{
	A vector of strings, each an id of a miogroups.
}
%\references{  }
\author{ Matthias Zschunke }

\keyword{ programming }% at least one, from doc/KEYWORDS
