\name{miotypes.read}
\alias{miotypes.read}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday private }
\description{
  Parse the miotypes section of the input file.
}
\usage{
miotypes.read(Lines)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Lines}{ The lines from the input file as given in the input file.}
}
\details{
  Do not use this in your code!
}
\value{
	A vector of strings, each a fully qualified class name of a MIOType.
}
%\references{ ~put references to the literature/web site here ~ }
\author{ Matthias Zschunke }

\keyword{ programming }% at least one, from doc/KEYWORDS
