% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCustom.R
\name{ModelCustom}
\alias{ModelCustom}
\alias{predict.ModelCustom}
\alias{summary.ModelCustom}
\title{ModelCustom}
\format{
an \link[R6:R6Class]{R6::R6Class} class inheriting from \link{Model}.
}
\usage{
\method{predict}{ModelCustom}(object, newdata, ...)

\method{summary}{ModelCustom}(object, ...)
}
\arguments{
\item{object}{a \link{ModelCustom} object}

\item{newdata}{a data.frame/data.table}

\item{...}{not being used.}
}
\value{
prediction

summary
}
\description{
A ModelCustom class.

print method.

an abstract method. Once implemented it should accept \code{newdata} as the first
argument and returns a \code{numeric()} vector or a \code{data.frame()} that contains
the predicted probabilities calculated using \code{self$params} and \code{newdata}.

summary method.
}
\section{Public fields}{

\itemize{
\item \code{params} :: named \code{numeric()}\cr
A named numeric vector containing parameter values of the model object.
\item \code{type} :: \code{character(1)}\cr
Type of the model.
\item \code{formula} :: \code{formula()}\cr
Model formula.
\item \code{terms} :: \code{character(1)}\cr
terms of the model. This gets generated using \code{stats::terms} on \code{formula}
}
}

