% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{Network}
\alias{Network}
\title{Network class}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Environment}<-\link{Entity}.
}
\description{
Network class. For creating a road network etc.
}
\note{
Network has a DataBackendDataTable object named "dodgr_graph" that basically
transform its attribute data into a \emph{dodgr} graph object. This is particularlly
useful for path finding applications etc.
}
\section{Construction}{
\preformatted{x <- Network$new()
}
\itemize{
\item .data\cr
A linestring \emph{sf} data.
\item id_col\cr
The name of the id column of \code{.data}.
\item from_node\cr
The name of the column that contains starting nodes.
\item to_node\cr
The name of the column that contains terminating nodes.
\item dist\cr
The name of the column that contains lengths or the distances of the edges.
}

Note that each row of the input data in \code{.data} must contain a linestring
or a multilinestring feature which basically an edge of a graph.
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{get_dodgr_graph(ids)}\cr
() -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a dodgr graph of the Network. This is just a shorthand for
\verb{Network$get_data(name = "dodgr_graph)}.
}
}

