% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{normalise_derived_vars}
\alias{normalise_derived_vars}
\title{Normalise the names of derived variables}
\usage{
normalise_derived_vars(.data)
}
\arguments{
\item{.data}{a data.frame object.}
}
\value{
a data.frame object.
}
\description{
Remove the \code{.} prefix from derived variables, if any, and return the data.frame.
}
\examples{

.df <- data.frame(a = rep(10, 10), b = rep(10, 10), .c = rep(10, 10), d.d = rep(10, 10))
normalise_derived_vars(.df)
}
\seealso{
omit_derived_vars
}
