% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{omit_derived_vars}
\alias{omit_derived_vars}
\title{Omit derived variables from a data.frame}
\usage{
omit_derived_vars(.data)
}
\arguments{
\item{.data}{a data.frame object.}
}
\value{
the input object with out columns that have \code{.} as prefix to their names.
}
\description{
In the attribute data of any inheritances of \link{Entity}, variables that have a
\code{.} prefix such as \code{.past_partner_id} is considered as a derived variable. It
is normally a variable that didn't belong to the original data that used to
populate the object in the first place. These derived variables usually get
created as the simulation evolves and past outcomes of entities get recorded.
}
\examples{

x <- data.frame(a = 1, b = 2, .c = 3)
omit_derived_vars(x)
}
\seealso{
normalise_derived_vars
}
