% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity-functions.R
\name{plot_history}
\alias{plot_history}
\title{Plot history data}
\usage{
plot_history(x, by_entity = TRUE)
}
\arguments{
\item{x}{a history data as a data.table or \link{Container}.}

\item{by_entity}{Subplots by entitiy. A logical value \code{TRUE} by default.}
}
\value{
a ggplot object
}
\description{
Plot history data
}
