% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC Curve}
\usage{
plot_roc_curve(
  roc.data,
  gene,
  aa_change = NULL,
  auc_roc_label = NULL,
  title = "Pathways Activation"
)
}
\description{
Plot ROC Curve
}
