% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-register.R
\name{pop_register}
\alias{pop_register}
\title{Population register}
\usage{
pop_register(x, ind_data, hh_data)
}
\arguments{
\item{x}{An Dymium's Agent object that contains Population object}

\item{ind_data}{a data.table contains individual data}

\item{hh_data}{a data.table contains household data}
}
\value{
a list contains ind_data and hh_data
}
\description{
Assign new ids to a new population. It is crucial that all ids in the relationship
columns do exist in the data of the new population.
}
