% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMultinomialLogit.R
\name{predict.ModelMultinomialLogit}
\alias{predict.ModelMultinomialLogit}
\title{predict method for ModelMultinomialLogit}
\usage{
\method{predict}{ModelMultinomialLogit}(object, newdata, chooser_id_col, choice_id_col, ...)
}
\arguments{
\item{object}{a \link{ModelMultinomialLogit} object}

\item{newdata}{a data.frame object}

\item{chooser_id_col}{a character that indicates the id column of the choosers in \code{newdata}.}

\item{choice_id_col}{a character that indicates the id column of the choices in \code{newdata}.}

\item{...}{not being used.}
}
\value{
a numeric vector
}
\description{
predict method for ModelMultinomialLogit
}
