% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_fastq_star_exp_sample}
\alias{preprocess_fastq_star_exp_sample}
\title{Process an expression sample from Oncobox FASTQ STAR pipeline into need format}
\usage{
preprocess_fastq_star_exp_sample(exp_sample_path, hsmart_ann, output_dir)
}
\arguments{
\item{exp_sample_path}{a str, specifying path to input expression sample}

\item{hsmart_ann}{a R object, containig HSMART ENSEMBL genes annotation}

\item{output_dir}{a str, specifying path to output dir}
}
\value{
a data frame with columns: \code{ensembl_gene_id}, \code{SYMBOL}, \verb{<sample_code>}
}
\description{
Process an expression sample from Oncobox FASTQ STAR pipeline into need format
}
