% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_fastq_star_samples_dir}
\alias{preprocess_fastq_star_samples_dir}
\title{Process in a given directory all expression samples from Oncobox FASTQ STAR pipeline into need format and saves output to dir}
\usage{
preprocess_fastq_star_samples_dir(
  input_exp_data_dir_path = "data/literature_validation_samples_data/exp_data/raw",
 
    output_exp_data_dir_path = "data/literature_validation_samples_data/exp_data/processed"
)
}
\description{
Process in a given directory all expression samples from Oncobox FASTQ STAR pipeline into need format and saves output to dir
}
