\name{probes.by.probelists}
\alias{probes.by.probelists}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ R for Mayday user function }
\description{
  Extract the probe ids contained in the probelists.
}
\usage{
probes.by.probelists(DATA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DATA}{ The DATA object. }
}
\details{
  Use this function to get a subset of all probe ids which are
  contained in the probelists.
}
\value{
	A vector of strings, each a probe id contained in one of the probelists.
}

%\references{ ~put references to the literature/web site here ~ }
\author{ Matthias Zschunke }

\keyword{ programming }% at least one, from doc/KEYWORDS
