% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{register}
\alias{register}
\title{Assign new ids to data}
\usage{
register(x, ..., only_primary_id_col = FALSE)
}
\arguments{
\item{x}{an \code{Entity} class object}

\item{...}{a list of data.frame objects to}

\item{only_primary_id_col}{:: \code{logical(1)}\cr
Default as \code{FALSE}. Only reassign ids for the primary id column of \code{...}.}
}
\value{
a data.frame object with new id values.
}
\description{
Assign new ids to data
}
