% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{rowSds}
\alias{rowSds}
\title{Row-wise standard deviations}
\usage{
rowSds(d)
}
\arguments{
\item{d}{dataframe or matrix}
}
\value{
numeric vector
}
\description{
Calculates standard deviation for every row in dataframe or matrix
}
\details{
Wraps sd() with apply()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{apply}} \code{\link{sd}}
}

