% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{rowVars}
\alias{rowVars}
\title{Row-wise variances}
\usage{
rowVars(d)
}
\arguments{
\item{d}{dataframe or matrix}
}
\value{
numeric vector
}
\description{
Calculates variance for every row in dataframe or matrix
}
\details{
Wraps var() with apply()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{apply}} \code{\link{var}}
}

