% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_loc}
\alias{run_loc}
\title{Detect whether the script is run locally or on server}
\usage{
run_loc()
}
\value{
Character vector
}
\description{
This function detects via username whether the script is run from a server or locally with the server system mounted.
}
\details{
This function basically returns a string with "~/server" if the user (as detected by Sys.info) is "malte" or "maltethodberg". This is useful when constructing file paths, since the same code will still get correct files if mounted via sshfs.
}
\author{
Malte Thodberg
}

