% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_choice}
\alias{sample_choice}
\title{Modified version of base::sample for programming in a Monte Carlo simulation}
\usage{
sample_choice(x, size = 1, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{a vector that contains value(s) that represents a choiceset.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\description{
This version of the sample function removes the feature where x is an integer of length 1
sampling via sample takes place from 1:x. This can bring unexpected behaviour
as mentioned in the documentation of \code{base::sample()}.
}
\examples{

sample_choice(7, 10) # equipvalent to rep(7, 10)
sample_choice(7, 1)
sample_choice(7) # which is equipvalent to the above
}
