% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choice.R
\name{sample_choice_table}
\alias{sample_choice_table}
\title{Sample a choice table}
\usage{
sample_choice_table(x, target = NULL, ...)
}
\arguments{
\item{x}{a \code{choice_table} object.}

\item{target}{a \link{Target} object.}

\item{...dots}{}
}
\value{
a character vector.
}
\description{
Sample a choice table
}
