% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{samples_names_to_case_ids}
\alias{samples_names_to_case_ids}
\title{Convert TCGA sample names into case IDs}
\usage{
samples_names_to_case_ids(sample_names)
}
\arguments{
\item{sample_names}{a vector, containing TCGA sample names}
}
\value{
a vector, contating TCGA case IDs
}
\description{
Convert TCGA sample names into case IDs
}
