% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenario.R
\name{set_active_scenario}
\alias{set_active_scenario}
\title{Set an active scenario directory.}
\usage{
set_active_scenario(name, .basedir = here::here())
}
\arguments{
\item{name}{name of the scenario folder to become active.}

\item{.basedir}{:: \code{character(1)}\cr
The base directory that the downloaded module will be saved at. \code{\link[here:here]{here::here()}} is
used to provide the default value which is is the root folder of the active RStudio project.}
}
\value{
a list contains scenario directories.
}
\description{
This is useful for event functions to access the current active scenario directory.
}
\examples{

\dontrun{
 use_module("test")
 set_active_scenario("test")
}
}
