% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choice.R
\name{simulate_choice.data.frame}
\alias{simulate_choice.data.frame}
\title{Simulate a choice situation}
\usage{
\method{simulate_choice}{data.frame}(probs, target = NULL)
}
\arguments{
\item{probs}{a data.table where the column names are choices and their values
are probabilities correspoding fors the choices. Each row represent the choice
probabilities of an agent.}

\item{target}{a Target object or a named list or \code{NULL}.}
}
\value{
a character vector of the same length as \code{prediction}.
}
\description{
Simulate a choice situation
}
\examples{

# TODO
}
