% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimulate.R
\name{simulate_choice}
\alias{simulate_choice}
\title{Simulate a choice situation}
\usage{
simulate_choice(probs, target)
}
\arguments{
\item{probs}{a data.table where the column names are choices and their values
are probabilities correspoding fors the choices. Each row represent the choice
probabilities of an agent.}
}
\value{
a character vector of the same length as \code{prediction}.
}
\description{
Simulate a choice situation
}
\examples{
TODO
}
