% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{source_data_to_matrix}
\alias{source_data_to_matrix}
\title{Combine source expression files into a matrix}
\usage{
source_data_to_matrix(
  exp_data_dir_path = "data/raw",
  output_matrix_path = "data/exp_matrix.txt",
  delim = ","
)
}
\description{
Combine source expression files into a matrix
}
