% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{SupportedTransitionModels}
\alias{SupportedTransitionModels}
\title{Classes of supported objects to be use in Transition.}
\usage{
SupportedTransitionModels()
}
\value{
a character vector
}
\description{
Currently, these classes are supported in the \code{model} argument of the Transition's
constructor: \link[caret:train]{caret::train}, a named \code{list}, and \link[data.table:data.table]{data.table::data.table}.
}
