% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Set a publication-ready theme for plots}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\description{
Set a publication-ready theme for plots
}
