% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_dwellings}
\alias{toy_dwellings}
\title{Toy microdata of dwellings}
\format{A data frame with 100 rows and 2 variables:
\describe{
\item{did}{(\code{integer(1)}) dwelling id}
\item{hid}{(\code{integer(1)}) household id}
\item{zid}{(\code{integer(1)}) zone id}
\item{structure}{(\code{character(1)}) structure of dwelling}
\item{bedroom}{(\code{integer(1)}) number of bedrooms}
}}
\usage{
toy_dwellings
}
\description{
This dataset contains a toy population of dwellings that is linked with \link{toy_individuals} and \link{toy_households}.
This dataset is only to be used for testing dymium packages.
}
\keyword{datasets}
