% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_households}
\alias{toy_households}
\title{Toy microdata of households}
\format{A data frame with 100 rows and 2 variables:
\describe{
\item{hid}{(\code{integer()}) household id}
\item{did}{(\code{integer()}) dwelling id}
}}
\usage{
data(toy_households)
}
\description{
This dataset contains a toy population of households that is linked with \link{toy_individuals}
that contains its corresponding toy population of individuals. This dataset is only to be used for testing
dymium packages.
}
\keyword{datasets}
