% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_individuals}
\alias{toy_individuals}
\title{Microdata of individuals}
\format{A data table with 346 rows and 9 variables:
\describe{
\item{pid}{(\code{integer()}) person id}
\item{hid}{(\code{integer()}) household id}
\item{age}{(\code{integer()}) age of the person}
\item{sex}{(\code{character()}) sex of the person - (male, female)}
\item{marital_status}{(\code{character()}) - (never married, married, divorced, widowed, not applicable)}
\item{partner_id}{(\code{integer()})}
\item{father_id}{(\code{integer()})}
\item{mother_id}{(\code{integer()})}
}}
\usage{
data(toy_individuals)
}
\description{
This dataset contains a toy population of individuals that is linkable with \link{toy_households}
that contains its corresponding toy population of households. This dataset is only to be used for testing
dymium packages.
}
\keyword{datasets}
