% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_transport_network}
\alias{toy_transport_network}
\title{Toy transport network data}
\format{
A \link[sf:sf]{sf::sf} data frame with 11 rows and 14 variables:
\describe{
\item{id}{(\code{integer()}) zone id}
\item{nb_lanes}{(\code{integer()})}
\item{speed}{(\code{double()})}
\item{capacity}{(\code{double()})}
\item{rd_type}{(\code{integer()})}
\item{fnode}{(\code{double()})}
\item{tnode}{(\code{double()})}
\item{geometry}{(\code{list()})}
\item{length}{(\code{double()})}
}
}
\usage{
toy_transport_network
}
\description{
Note that this dataset requires the \code{sf} package to work properly. This
dataset contains street network within some SA2 zones of Melborune CBD,
Australia in a \link[sf:sf]{sf::sf} format. This network is contained within the boundaries
of \link{toy_zones}.
}
\examples{
toy_transport_network
}
\seealso{
toy_zones
}
\keyword{datasets}
