% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_zones}
\alias{toy_zones}
\title{Toy zone data}
\format{A \link[sf:sf]{sf::sf} data frame with 11 rows and 14 variables:
\describe{
\item{zid}{(\code{integer()}) zone id}
\item{sa2_main}{(\code{integer()})}
\item{sa2_5dig11}{(\code{character()})}
\item{sa2_name11}{(\code{character()})}
\item{sa3_code11}{(\code{character()})}
\item{sa3_name11}{(\code{character()})}
\item{sa4_code11}{(\code{character()})}
\item{sa4_name11}{(\code{character()})}
\item{gcc_code11}{(\code{character()})}
\item{gcc_name11}{(\code{character()})}
\item{ste_code11}{(\code{character()})}
\item{ste_name11}{(\code{integer()})}
\item{albers_sqm}{(\code{double()})}
\item{geometry}{(\code{list()})}
}}
\usage{
toy_zones
}
\description{
Note that this dataset requires the \code{sf} package to work properly. This
dataset contains 11 SA2 zones of Melborune CBD, Australia in a \link[sf:sf]{sf::sf} format.
}
\keyword{datasets}
