% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{transition}
\alias{transition}
\title{Simulate a transition of entities}
\usage{
transition(
  entity,
  model,
  target = NULL,
  targeted_agents = NULL,
  update_attr = NULL
)
}
\arguments{
\item{entity}{an \link{Entity} object}

\item{model}{a model object that belongs to the classes in \link{SupportedTransitionModels}.}

\item{target}{a named list that is the target for alignment.}

\item{update}{default as NULL. This indicates whether \code{entity} should be updated
using the outcomes from the transtion. To update an attribute of \code{entity} the name
of the attribute to be updated must be specified as a character value.}
}
\value{
a data.table with two columns: id and response.
}
\description{
Simulate a transition of entities
}
\examples{

# load data
create_toy_population()
Ind <- pop$get("Individual")

# create model
model_lm <- glm(age ~ sex + marital_status, data = Ind$get_data(), family = "gaussian")
model_glm <- glm(I(sex == "male") ~ age + marital_status, data = Ind$get_data(), family = "binomial")

# simulation transition
transition(Ind, model_lm)
transition(Ind, model_glm)
}
