% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionRegression.R
\name{TransitionRegression}
\alias{TransitionRegression}
\title{Transition for regression model}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Transition}}
\description{
Note that, TransitionRegression only returns a raw output of the simulation result.
}
\section{Construction}{
\preformatted{Transition$new(x, model, target = NULL, targeted_agents = NULL)
}

See \link{Transition}, for a description of the arguments.
}

\seealso{
\link{TransitionClassification}
}
