% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Unnest data.table}
\usage{
unnest_dt(dt, cols)
}
\arguments{
\item{dt}{:: (\code{data.frame()})\cr
a data.frame object.}

\item{cols}{:: (\code{character()})\cr
a character vector denoting columns to be unnested.}
}
\value{
a data.table
}
\description{
!! Note that this is only working when all list columns in the given
data.frame are to be unnested. See https://github.com/dymium-org/dymiumCore/issues/79.
}
