% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_module}
\alias{use_module}
\title{Create and setup a module folder.}
\usage{
use_module(name)
}
\arguments{
\item{name}{Name of the module.}
}
\description{
This function creates a new module inside the modules folder of an active r project.
If the 'modules' folder doesn't exist it will create it then adds a new folder
with the name as specified in the \code{name} argument inside the 'modules' folder.
R scripts to be used across the module will be added which contain the following:
\itemize{
\item a lgr logger script,
\item a script that contains constant values, and
\item a script for storing helper functions.
}

Since dymium modules use the 'modules' and 'checkmate' packages, if these
packages are not installed the function will ask whether you would like to
install them or not.

Note that, to add event functions to a module see \link{use_event}.
}
\examples{
\dontrun{
# Note: running this will create a folder called "modules" and a sub-folder
#       to your working directory within the folder called "demography"
use_module(name = "demography")
}
}
