% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_module_readme}
\alias{use_module_readme}
\title{Add a README rmarkdown file to an existing module}
\usage{
use_module_readme(name)
}
\arguments{
\item{name}{name of an existing module}
}
\description{
Add a README rmarkdown file to an existing module
}
\examples{
\dontrun{
# this assumes that you have a module named 'demography'
use_module_readme(name = "demography")
}
}
