% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_scenario}
\alias{use_scenario}
\title{Create a dymium scenario folder}
\usage{
use_scenario(name, active = TRUE)
}
\arguments{
\item{name}{Name of the scenario}

\item{active}{a logical value with defauly being FALSE. This determines whether
to set the current active scenario to this newly created scenario or not.}
}
\description{
Creates a scenario folder using a standard format which contains an 'inputs' folder,
an 'outputs' folder inside.
}
\examples{

\dontrun{
  use_scenario(name = "demography")
}
}
