% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_linkages}
\alias{validate_linkages}
\title{Validate the linkages between Entities in World}
\usage{
validate_linkages(world)
}
\arguments{
\item{world}{a \link{World} object.}
}
\value{
a logical value
}
\description{
Validate the linkages between Entities in World
}
\examples{

create_toy_world()

validate_linkages(world)
}
