% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_max_n}
\alias{which_max_n}
\alias{which_min_n}
\title{Return the indices of n maximum or minimum values}
\usage{
which_max_n(x, n = 1)

which_min_n(x, n = 1)
}
\arguments{
\item{x}{:: (\code{numeric()})\cr
an numeric vector}

\item{n}{:: (\code{integer(1)})\cr
number of values to return.}
}
\value{
an \code{integer()} vector.
}
\description{
Return the indices of n maximum or minimum values
}
\examples{

which_max_n(1:4, 2)
which_min_n(1:4, 2)

}
