% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agent.R
\name{Agent}
\alias{Agent}
\title{Agent}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Entity}.}
\description{
Agent class is the generic class for creating urban agents such
as individuals, households, firms, dwelling and more.
}
\section{Construction}{
\preformatted{Agt <- Agent$new(...)
}
}

\section{Fields}{

\itemize{
\item \code{data}::\code{data.table::data.table()}\cr
Contains the attributes of agent. One of the attributes must be a unique id field
}
}

\section{Methods}{

\itemize{
\item \strong{DEPRECATING} \code{initialise_data(data, id_col)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}, \code{character(1)}) -> \code{NULL}\cr
Store data to data while making sure that \code{data} meets all the
requirements. \code{data}, which is also the first set of data for the agent object,
will be stored at the first index in \code{private$.data}, a named list of data objects, and named
as \code{attrs}.
\item \code{add_new_agents(parent_ids = NULL, data = NULL)}\cr
(\code{integer()}|\code{character()}, \code{\link[data.table:data.table]{data.table::data.table()}}) -> \code{NULL}\cr
Adds new agent records to data.
\item \code{get_latest_agent_id(ids)}\cr
(\code{integer()}) -> \code{integer()}\cr
Return the id of the last agent that was created.
\item \code{get_new_agent_ids()}\cr
() -> \code{integer()}\cr
Return the ids of new agents.
\item \code{is_alive(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Check if the agents with id in \code{ids} are still alive/active in its data object.
or not.
\item \code{generate_new_ids(n)}\cr
(\code{integer(1)}) -> \code{integer()}\cr
Generates \code{n} number of new ids.
\item \code{subset_ids(expression)}\cr
(\code{expression}) -> \verb{[data.table::data.table()]}\cr
Return ids of rows matches the expression.
\item \code{hatch(ids)}\cr
(\code{integer()})\cr
Clone existing agents in \code{ids} and inherit all attributes from their parents
except for the id attribute.
\item \code{add(.data)}\cr
Add new agents in \code{.data} to the existing pool of agents. It should be noted,
that all new agents must have all the columns that the existing agents possess
with the exception of derived variables which is denoted by the \code{.} prefix.
For example, \code{.past_partner_id} is a derived variable which is allowed to be
missing in the new agents' data.
}
}

