% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asset.R
\name{Asset}
\alias{Asset}
\title{Asset class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Entity}.}
\description{
An \link{Asset} represents physical objects such as buildings and vehicles that
can be owned/rent/belong by or to agents of the \link{Agent} sub classes.
For example, \link{Household} agents own or rent residential buildings.
}
\section{Construction}{
\preformatted{x <- Entity$new(databackend, .data, id_col)
}

Stores \code{.data} as a DataBackend object inside the object's list of data (\code{private$.data})
and registers the \code{id_col} (\code{private$.id_col}).
\itemize{
\item \code{databackend} :: \link{DataBackend} classes\cr
An \link[R6:R6Class]{R6::R6Class} generator that inherits from \code{DataBackend}.
\item \code{.data} :: \code{data.frame}\cr
A object that inherits from \code{data.frame}.
\item \code{id_col} :: \code{character(1)}\cr
The id column of \code{.data}.
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{owner(ids)}\cr
(\code{integer()}) -> \code{integer()}\cr
Returns owner ids.
}
}

