% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{bioc_packages}
\alias{bioc_packages}
\title{Bioconductor Packages}
\format{Character vector}
\usage{
bioc_packages
}
\description{
Names of biconductor packages that should always be loaded.
}
\details{
The list of packages includes:

Installer: BiocInstaller

Genomic Arithmetic: Biostrings, IRanges, GenomicRanges, rtracklayer

Differential Expression: limma, edgeR, DESeq2
}
\author{
Malte Thodberg
}
\keyword{datasets}

