% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{Building}
\alias{Building}
\title{Building class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Create Building class, extended Agent class.
}
\section{Construction}{
\preformatted{x <- Entity$new(databackend, .data, id_col)
}

Stores \code{.data} as a DataBackend object inside the object's list of data (\code{private$.data})
and registers the \code{id_col} (\code{private$.id_col}).
\itemize{
\item \code{databackend} :: a \link[R6:R6Class]{R6::R6Class} generator\cr
A \link[R6:R6Class]{R6::R6Class} generator that inherits from \code{DataBackend}.
\item \code{.data} :: \code{data.frame()}\cr
A object that inherits from \code{data.frame}.
\item \code{id_col} :: \code{character(1)}\cr
The id column of \code{.data}.
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item \code{is_occupied(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is occupied
\item \code{is_vacant(ids)}\cr
(\code{integer()}) -> \code{logical()}\cr
Returns TRUE if dwelling in ids is vacant.
}
}

