% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{Building}
\alias{Building}
\title{Building class}
\description{
Create a building object.\preformatted{x <- Building$new(.data, id_col, owner)
}

Stores \code{.data} as a DataBackend object inside the object's list of data (\code{private$.data})
and registers the \code{id_col} (\code{private$.id_col}).
\itemize{
\item \code{.data} :: \code{data.frame}\cr
A object that inherits from \code{data.frame}.
\item \code{id_col} :: \code{character(1)}\cr
The id column of \code{.data}.
\item \code{owner} :: an \link[R6:R6Class]{R6::R6Class} object that inherits \link{Agent}.\cr
}
}
\section{Fields}{


\code{NULL}\cr
}

\section{Methods}{

\itemize{
\item \code{is_vacant(ids)}\cr
(\code{integer()}) -> (\code{logical()})\cr
Check if the assets in ids are vacant
\item \code{vacate(ids)}\cr
(\code{integer()})\cr
This removes the owners of the assets in \code{ids}.
}
}

