% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_entity}
\alias{check_entity}
\alias{assert_entity}
\alias{test_entity}
\alias{expect_entity}
\title{Check if an argument is from the Entity class.}
\usage{
check_entity(x)

assert_entity(x, .var.name = checkmate::vname(x), add = NULL)

test_entity(x)

expect_entity(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{an object to be checked.}
}
\value{
a logical value if TRUE or an character containing an error message.
}
\description{
Check if an argument is from the Entity class.
}
