% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_entity_ids}
\alias{check_entity_ids}
\alias{test_entity_ids}
\alias{expect_entity_ids}
\alias{assert_entity_ids}
\title{Check if ids exist in Entity.}
\usage{
check_entity_ids(x, ids, include_removed_data = FALSE, informative = TRUE)

test_entity_ids(x, ids, include_removed_data = FALSE, informative = TRUE)

expect_entity_ids(
  x,
  ids,
  include_removed_data = FALSE,
  informative = TRUE,
  info = NULL,
  label = vname(x)
)

assert_entity_ids(
  x,
  ids,
  include_removed_data = FALSE,
  informative = TRUE,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{an R6 \link{Entity} object}

\item{ids}{\link{\code{integer()}}\cr
Ids of Entity to check.}

\item{include_removed_data}{\link{\code{logical(1)}}\cr
Should the removed data of \code{x} be checked.}

\item{informative}{\link{\code{logical(1)}}\cr
Whether to return the missing ids in error message.}
}
\value{
a logical value.
}
\description{
Check if ids exist in Entity.
}
