% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_entity_ids}
\alias{check_entity_ids}
\alias{test_entity_ids}
\alias{expect_entity_ids}
\alias{assert_entity_ids}
\title{Check if ids exist in Entity.}
\usage{
check_entity_ids(x, ids, informative = TRUE)

test_entity_ids(x, ids, informative = TRUE)

expect_entity_ids(x, ids, informative = TRUE, info = NULL, label = vname(x))

assert_entity_ids(
  x,
  ids,
  informative = TRUE,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{an R6 \link{Entity} object}

\item{ids}{\code{\link[=integer]{integer()}}\cr
Ids of Entity to check.}

\item{informative}{\link{logical(1)}\cr
Whether to return the missing ids in error message.}
}
\value{
a logical value.
}
\description{
Check if ids exist in Entity.
}
