% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_required_models}
\alias{check_required_models}
\alias{assert_required_models}
\alias{test_required_models}
\alias{expect_required_models}
\title{Check if an argument x has all the required models}
\usage{
check_required_models(x, names, check_supported_model = TRUE)

assert_required_models(
  x,
  names,
  check_supported_model = TRUE,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_required_models(x, names, check_supported_model = TRUE)

expect_required_models(
  x,
  names,
  check_supported_model = TRUE,
  info = NULL,
  label = vname(x)
)
}
\arguments{
\item{x}{a named list.}

\item{names}{a character vector contains names of the required models}

\item{check_supported_model}{Whether to check that each element in \code{x} are
supported by \link{Transition} using \code{assert_transition_supported_model()}.}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{info}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{label}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}
}
\value{
TRUE if all the required models exist else throws an error.
}
\description{
Check if an argument x has all the required models
}
