% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_transition_supported_model}
\alias{check_transition_supported_model}
\alias{assert_transition_supported_model}
\alias{test_transition_supported_model}
\alias{expect_transition_supported_model}
\title{Check if an argument is an object that is supported by \link{Trans}}
\usage{
check_transition_supported_model(x)

assert_transition_supported_model(
  x,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_transition_supported_model(x)

expect_transition_supported_model(x, info = NULL, label = vname(x))
}
\arguments{
\item{x}{(\code{any})\cr
an object to check.}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{add}{\link[checkmate:AssertCollection]{checkmate::AssertCollection}\cr
Collection to store assertions. See \link[checkmate:AssertCollection]{checkmate::AssertCollection}.}

\item{info}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{label}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}
}
\value{
a logical value if TRUE, if false then returns an error message
}
\description{
Check if an argument is an object that is supported by \link{Trans}
}
