% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{code_packages}
\alias{code_packages}
\title{Development Packages}
\usage{
code_packages()
}
\value{
Character vector
}
\description{
Simply return a vector with names of packages used for development
}
\details{
The list of packages includes:

Development: Rccp, devtools, roxygen2, pryr, profr

RStudio: rstudioapi, manipulate

Terminal: setwidth, colorout
}
\author{
Malte Thodberg
}

