% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContainerGeneric.R
\name{ContainerGeneric}
\alias{ContainerGeneric}
\title{ContainerGeneric}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[dymiumCore:Generic]{dymiumCore::Generic}
}
\description{
ContainerGeneric is the base class for the \link{Container} classes. \link{ContainerGeneric}
isn't supposed to be use directly.
}
\section{Construction}{
\preformatted{x <- ContainerGeneric$new()
}
\itemize{
\item NULL\cr
}
}

\section{Public Fields}{

\itemize{
\item \code{Cont}\cr
A named list that contains objects of any type.
}
}

\section{Public Methods}{

\itemize{
\item \code{add(x)}\cr
(any objects)\cr
Add \code{x} to self$Cont.
\item \code{check_pos(x)}\cr
(\code{integer(1)}) -> (\code{logical(1)})\cr
Check if the position in \code{x} exists in \code{self$Cont}.
\item \code{get(x)}\cr
(\code{integer(1)}) -> (an object)\cr
Get a object in \code{x} element of \code{self$Cont}.
\item \code{n()}\cr
() -> (\code{integer(1)})\cr
Returns the length of \code{self$Cont}.
\item \code{print()}\cr
\item \code{remove(x)}\cr
(\code{integer(1)})\cr
Remove the object at the \code{x} element of \code{self$Cont}.
\item \code{unpack(target)}\cr
(\link[dymiumCore:Container]{dymiumCore::Container})\cr
Add all the objects in \code{self$Cont} to another \code{target} Container.
\item \code{summary()}\cr
Abstract method.
}
}

