\name{CoRe.empiricalOdds}
\alias{CoRe.empiricalOdds}
\title{Empirical odds of number of fitness genes per number of cell lines}
\usage{
CoRe.empiricalOdds(observedCumSum,
                   simulatedCumSum)
}
\arguments{
\item{observedCumSum}{Observed profile of cumulative sum of numbers of fitness genes in fixed number of cell lines. This is generated by the \code{CoRe.panessprofile} function.}

\item{simulatedCumSum}{Random profiles of cumulative sum of fitness genes in fixed number of cell lines. This is generated by the function \code{CoRe.generateNullModel}.}
}
\value{
A named vector:
\item{odds}{log base 10 odd ratios of observed versus expected cumulative sums of number of fitness genes across fixed numbers of cell lines.}
}
\description{
This function calculates log10 odd ratios of observed vs expected profiles of cumulative number of fitness genes in fixed number of cell lines.
}
\details{
This function is used by the AdAM method [1] to calculate odd ratios of observed vs expected profiles of cumulative number of fitness genes in fixed number of cell lines. Expected values are the mean of those observed across randomised version of the initial binary matrix. This function is used in cascade after calls to \code{CoRe.generateNullModel} and \code{CoRe.panessprofile}.
}
\examples{
## Downloading dependency matrix
## for > 300 cancer cell lines from [1,2]
BinDepMat<-CoRe.download_BinaryDepMatrix()

## Extracting dependency submatrix for
## Non-Small Cell Lung Carcinoma cell lines only
LungDepMap<-CoRe.extract_tissueType_SubMatrix(BinDepMat)

## Computing number of fitness genes across fixed numbers
## of cell lines and its cumulative sums
observed <- CoRe.panessprofile(depMat = LungDepMap)

## Simulating Null model for the number of fitness genes
## across numbers of cell lines and their cumulative sums
null_m<-CoRe.generateNullModel(depMat = LungDepMap)

## Computing empirical odds of number of fitness genes per number of cell lines
logOdds <- CoRe.empiricalOdds(observedCumSum=observed$CUMsums,simulatedCumSum=null_m$nullCumSUM)
logOdds
}
\seealso{
CoRe.panessprofile, CoRe.generateNullModel
}
\references{
[1] Behan FM, Iorio F, Picco G, Gonçalves E, Beaver CM, Migliardi G, et al. Prioritization of cancer therapeutic targets using CRISPR-Cas9 screens. Nature. 2019;568:511–6.

[2] Dwane L, Behan FM, Gonçalves E, et al. Project Score database: a resource for investigating cancer cell dependencies and prioritizing therapeutic targets. Nucleic Acids Res. 2021 Jan 8;49(D1):D1365-D1372.
}
\author{
C. Pacini, E. Karakoc, A. Vinceti & F. Iorio
}
\keyword{functions}
