% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{core_packages}
\alias{core_packages}
\title{Common Packages}
\format{Character vector}
\usage{
core_packages
}
\description{
Names of basic packages that should always be loaded.
}
\details{
The list of packages includes:

Data manipulation: magrittr, readr, tidyr, dplyr

Special data formats: stringr, lubridate

Plotting: grid, gridExtra, ggplot2, GGally, ggthemes, ggExtra, ggrepel, RColorBrewer, VennDiagram, pheatmap, wesanderson

Performance: matrixStats, parallel
}
\author{
Malte Thodberg
}
\keyword{datasets}

