% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{core_packages}
\alias{core_packages}
\title{Common Packages}
\usage{
core_packages()
}
\value{
Character vector
}
\description{
Simply return a vector with names of most commen packages
}
\details{
The list of packages includes:

Data manipulation: magrittr, readr, tidyr, dplyr

Special data formats: stringr, lubridate

Plotting: grid, gridExtra, ggplot2, GGally, ggthemes, ggExtra, RColorBrewer, VennDiagram, pheatmap

Performance: parallel
}
\author{
Malte Thodberg
}

