% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataFrame.R
\name{DataBackendDataFrame}
\alias{DataBackendDataFrame}
\title{data.frame as data backend}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{DataBackend}.}
\description{
Data backend of data.frame
}
