% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataTable}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
SpatialData class is immutable.
}
\section{Construction}{
\preformatted{x <- SpatialData$new(.data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
The input \code{\link[data.table:data.table]{data.table::data.table()}}.
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr
}
}

\section{Methods}{

\itemize{
\item \code{add(.data)}\cr
(\link[data.table:data.table]{data.table::data.table}) -> \code{NULL}\cr
Add data.
}
}

\section{Super classes}{
\code{\link[dymiumCore:Generic]{dymiumCore::Generic}} -> \code{\link[dymiumCore:DataBackend]{dymiumCore::DataBackend}} -> \code{\link[dymiumCore:DataBackendDataFrame]{dymiumCore::DataBackendDataFrame}} -> \code{DataBackendDataTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataBackendDataTable$new()}}
\item \href{#method-add}{\code{DataBackendDataTable$add()}}
\item \href{#method-remove}{\code{DataBackendDataTable$remove()}}
\item \href{#method-clone}{\code{DataBackendDataTable$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="debug">}\href{../../dymiumCore/html/Generic.html#method-debug}{\code{dymiumCore::Generic$debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="is_dymium_class">}\href{../../dymiumCore/html/Generic.html#method-is_dymium_class}{\code{dymiumCore::Generic$is_dymium_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="message">}\href{../../dymiumCore/html/Generic.html#method-message}{\code{dymiumCore::Generic$message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_debug">}\href{../../dymiumCore/html/Generic.html#method-msg_debug}{\code{dymiumCore::Generic$msg_debug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_info">}\href{../../dymiumCore/html/Generic.html#method-msg_info}{\code{dymiumCore::Generic$msg_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_trace">}\href{../../dymiumCore/html/Generic.html#method-msg_trace}{\code{dymiumCore::Generic$msg_trace()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="Generic" data-id="msg_warn">}\href{../../dymiumCore/html/Generic.html#method-msg_warn}{\code{dymiumCore::Generic$msg_warn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackend" data-id="class">}\href{../../dymiumCore/html/DataBackend.html#method-class}{\code{dymiumCore::DataBackend$class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="colnames">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-colnames}{\code{dymiumCore::DataBackendDataFrame$colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="get">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-get}{\code{dymiumCore::DataBackendDataFrame$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="get_removed">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-get_removed}{\code{dymiumCore::DataBackendDataFrame$get_removed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="head">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-head}{\code{dymiumCore::DataBackendDataFrame$head()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="n">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-n}{\code{dymiumCore::DataBackendDataFrame$n()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="ncol">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-ncol}{\code{dymiumCore::DataBackendDataFrame$ncol()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="nrow">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-nrow}{\code{dymiumCore::DataBackendDataFrame$nrow()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dymiumCore" data-topic="DataBackendDataFrame" data-id="view">}\href{../../dymiumCore/html/DataBackendDataFrame.html#method-view}{\code{dymiumCore::DataBackendDataFrame$view()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendDataTable$new(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendDataTable$add(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendDataTable$remove(rows, cols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendDataTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
