% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataTable}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
SpatialData class is immutable.
}
\section{Construction}{
\preformatted{x <- SpatialData$new(.data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
The input \code{\link[data.table:data.table]{data.table::data.table()}}.
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr
}
}

\section{Methods}{

\itemize{
\item \code{add(.data)}\cr
(\link[data.table:data.table]{data.table::data.table}) -> \code{NULL}\cr
Add data.
}
}

