% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-event.R
\name{dm_get_model}
\alias{dm_get_model}
\title{Get model objects from \link{World}.}
\usage{
dm_get_model(x, model_names)
}
\arguments{
\item{x}{a \link{World} object.}

\item{model_names}{a character vector of the named \link{Model} objects to be retrieved.}
}
\value{
a named \code{list} containing \link{Model} objects
}
\description{
Get model objects from \link{World}.
}
\examples{

create_toy_world()
testModel <- list(a = 1)
world$add(testModel, "testModel")
get_models(world, "testModel")
}
