% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-utils.R
\name{dm_save}
\alias{dm_save}
\title{save dymium object}
\usage{
dm_save(x, path = getOption("dymium.output_dir"), .overwrite = FALSE)
}
\arguments{
\item{x}{a dymium class object}

\item{path}{a character, containing a path. This is only needed if the object
doesn't have a config path. Otherwise it will be save in the standard dymium
output path, which is \verb{<config-path>/outputs/iter-<number>}.}

\item{.overwrite}{\code{logical(1)}}
}
\value{
the input `object`` invisibly
}
\description{
This function save the input object and returns the object. This is designed to
work in a dymium pipeline. Note that this function can overwrite saved objects
but not existing folders.
}
