% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dymiumCore-package.R
\docType{package}
\name{dymiumCore}
\alias{dymiumCore}
\title{\code{dymiumCore} package}
\description{
Dynamic microsimulation framework for integrated urban models.
}
\details{
See the README on
\href{https://github.com/dymium-org/dymiumCore#readme}{GitHub}
}
\keyword{internal}
