% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{Environment}
\alias{Environment}
\title{Environment class}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Entity}.
}
\description{
The base class for physical environment, infrastructure and boundaries. This
is different from Asset as Environment has no owner or belongs to the public.
}
\section{Construction}{
\preformatted{x <- Environment$new()
}
\itemize{
\item NULL\cr
}
}

\section{Fields}{

\itemize{
\item \code{NULL}\cr`
}
}

\section{Methods}{

\itemize{
\item NULL
}
}

