% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{Generic}
\alias{Generic}
\title{Generic class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Can be used as a template for other objects. This contains useful utility
functions for abstraction objects. See the public members (methods).
}
\section{Construction}{
\preformatted{Obj <- Generic$new(...)
}
}

\section{Fields}{

\itemize{
\item \code{NULL}
}
}

\section{Methods}{

\itemize{
\item \code{debug()}\cr
() -> \code{self & private}\cr
Makes debugging the object easier by assigning self and private to the
global environment of the current seesion.
\item \code{message(x, ...)}\cr
Print \code{message} onto the console with the object classname append in front of
the message. To turn this off, set options(dymium.agent.verbose = FALSE).
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Generic$new()}}
\item \href{#method-debug}{\code{Generic$debug()}}
\item \href{#method-msg_trace}{\code{Generic$msg_trace()}}
\item \href{#method-msg_debug}{\code{Generic$msg_debug()}}
\item \href{#method-msg_info}{\code{Generic$msg_info()}}
\item \href{#method-msg_warn}{\code{Generic$msg_warn()}}
\item \href{#method-message}{\code{Generic$message()}}
\item \href{#method-is_dymium_class}{\code{Generic$is_dymium_class()}}
\item \href{#method-class}{\code{Generic$class()}}
\item \href{#method-clone}{\code{Generic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug"></a>}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$debug()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msg_trace"></a>}}
\subsection{Method \code{msg_trace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$msg_trace(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msg_debug"></a>}}
\subsection{Method \code{msg_debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$msg_debug(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msg_info"></a>}}
\subsection{Method \code{msg_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$msg_info(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-msg_warn"></a>}}
\subsection{Method \code{msg_warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$msg_warn(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-message"></a>}}
\subsection{Method \code{message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$message(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_dymium_class"></a>}}
\subsection{Method \code{is_dymium_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$is_dymium_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-class"></a>}}
\subsection{Method \code{class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
