% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{Generic}
\alias{Generic}
\title{Generic class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Can be used as a template for other objects. This contains useful utility
functions for abstraction objects. See the public members (methods).
}
\section{Construction}{
\preformatted{Obj <- Generic$new(...)
}
}

\section{Fields}{

\itemize{
\item \code{NULL}
}
}

\section{Methods}{

\itemize{
\item \code{debug()}\cr
() -> \code{self & private}\cr
Makes debugging the object easier by assigning self and private to the
global environment of the current seesion.
\item \code{log(desc, value, tag = "", time = .get_sim_time(), .lg = parent.frame()[["lg"]])}\cr
(\code{character(1)}, any objects, \code{character(1)}, \code{integer(1)}, \link[lgr:Logger]{lgr::Logger})\cr
For logging simulation outcomes to \code{private$.log}. \code{get_log(x)} can be used to
extract the log data from any object that inherit \code{Generic}. Note that, the
logger in \code{log} can be specified. By default, it tries to look in the \code{parent.frame}
the calling scope to find a logger named \code{lg} which is the default name used in
dymium modules. If not found, then it will automatically use dymiumCore's logger.
\item \code{is_dymium_class()}\cr
Returns \code{TRUE}. Use for internal checking.
}
}

