% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-event.R
\name{get_models}
\alias{get_models}
\title{Get model objects from \link{World}.}
\usage{
get_models(x, model_names, as_r6model = FALSE)
}
\arguments{
\item{x}{a \link{World} object.}

\item{model_names}{a character vector of the named \link{Model} objects to be retrieved.}

\item{as_r6model}{default as FALSE, if TRUE all \code{required_models} will be returned
as \link{Model} objects.}
}
\value{
a named \code{list} containing \link{Model} objects
}
\description{
Get model objects from \link{World}.
}
\examples{

create_toy_world()
testModel <- list(a = 1)
world$add(testModel, "testModel")
get_models(world, "testModel")
}
