% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{h}
\alias{h}
\title{Short head}
\usage{
h(x, ...)
}
\value{
Similar to head()
}
\description{
Shorter call to head
}
\details{
Similar to head()
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{head}}
}

