% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/History.R
\name{History}
\alias{History}
\title{History class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Generic}}
\description{
Create History class, record life-course changes of agents.
}
\section{Construction}{
\preformatted{Hist <- History$new()
}
}

\section{Public Fields}{

\itemize{
\item NULL
}
}

\section{Public Methods}{

\itemize{
\item \code{add(ids, event, time = .get_sim_time())}\cr
(\code{integer()}, \code{character(1)}, \code{integer(1)})\cr
Record occurences of an event occured to entities in \code{ids}.
\item \code{count(ids, event)}\cr
(\code{integer()}, \code{character(1)})\cr
count the nu
}
}

\section{Public methods}{

\code{add(id, time, event)} -
add changes (\code{event}) of agents (\code{id}) with a timestamp (\code{time})

\code{count(ids = NULL, event = NULL)} -
Returns the number of occurence (N) of events for each id.

\code{inspect(ids)} -
Return lifecourse history of the agents in ids by time.

\code{impute(data, id_col, event = NULL)} -
Return the original data imputed with the number of occurence
of the event specified in \code{event} arg. \code{id_col} is used to
merge the output of \code{self$count()} with data.

\code{debug()} -
Makes debugging the object easier by assigning self and
private to the global environment of the current seesion.

\code{get_data(copy = TRUE)} -
return a copy of \code{private$data}. Since \code{private$data} is
a data.table object, if \code{copy} is set to FALSE a reference
to \code{private$data} will be returned.

\code{message()} -
Print \code{message} onto the console with the object name
append in front of the message.
}

\section{Private methods}{

\code{init_data()} -
Initialise data in the correct structure.

\code{data_template()} -
Create a data template which mimics the structure and
column types of private$data.

\strong{NOTE}: event_as_list can take up a lot of memory since
all event entries will be converted to lists.
}

\examples{

h$add(
  id = sample(1L:10L, 20, replace = TRUE),
  time = sample(1L:4L, 20, replace = TRUE),
  event = sample(c("Birth", "LeaveHome", "Marriage"), 20, replace = TRUE))
}
