% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-class.R
\name{is_dymium_class}
\alias{is_dymium_class}
\title{is dymium class}
\usage{
is_dymium_class(x)
}
\arguments{
\item{x}{any object}
}
\value{
a logical value
}
\description{
is dymium class
}
\examples{

# Not Dymium object
library(R6)
x <- R6Class()
is_dymium_class(x)

# Dymium object
Pop <- Population$new(
  ind_data = toy_individuals,
  hh_data = toy_households,
  pid_col = "pid",
  hid_col = "hid"
)
is_dymium_class(Pop)
}
