% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting \link{Generic}.}
\description{
Provides an interface for model objects to be use in \link{Transition} and \link{World}.
}
\section{Construction}{
\preformatted{x <- Model$new(x)
}
\itemize{
\item \code{x} :: (\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
A model object that compatible.
}
}

\section{Public Methods}{

\itemize{
\item \code{get()}\cr
() -> (\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
Get a model object.
\item \code{set()}\cr
() -> (\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
Get a model object.
\item \code{modify()}\cr
\item \code{simulate()}\cr
\item \code{print()}\cr
}
}

