% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model class}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting \link{Generic}.}
\description{
A container for the supported model objects (see \link{SupportedTransitionModels}).
}
\section{Construction}{
\preformatted{x <- Model$new(x)
}
\itemize{
\item \code{x} :: (\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
A model object that compatible.
}
}

\section{Public Methods}{

\itemize{
\item \code{get()}\cr
() -> (\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
Get a model object.
\item \code{set(x)}\cr
(\link[caret:train]{caret::train} | \link[data.table:data.table]{data.table::data.table} | named \code{list})\cr
Get a model object.
\item \code{modify()}\cr
An abstract method.
\item \code{simulate()}\cr
An abstract method.
\item \code{print()}\cr
}
}

\examples{

world <- World$new()

prob_model <- list(yes = 0.95, no = 0.05)

world$add(x = prob_model, name = "prob_model")

world$get(x = "prob_model")

world$get_model(x = "prob_model")

}
