% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCustom.R
\name{ModelCustom}
\alias{ModelCustom}
\alias{predict.ModelCustom}
\alias{summary.ModelCustom}
\title{ModelCustom}
\usage{
\method{predict}{ModelCustom}(object, newdata)

\method{summary}{ModelCustom}(x)
}
\arguments{
\item{object}{a \link{ModelCustom} object}

\item{newdata}{a data.frame/data.table}

\item{x}{a \link{ModelCustom} object}

\item{params}{a named \code{numeric()}.}

\item{formula}{a model \code{formula()}.}

\item{type}{type of the model.}

\item{preprocessing_fn}{a pre-processing function that gets applied to the
data given to the \code{predict} method before making the prediction.}
}
\value{
\code{NULL}

prediction

summary
}
\description{
A ModelCustom class.

Constructor function.

print method.

an abstract method. Once implemented it should accept \code{newdata} as the first
argument and returns a \code{numeric()} vector or a \code{data.frame()} that contains
the predicted probabilities calculated using \code{self$params} and \code{newdata}.

summary method.
}
\section{Fields}{

\describe{
\item{\code{params}}{named \code{numeric()}\cr
a named numerical vector containing parameter values of the model object.}

\item{\code{type}}{\code{character(1)}\cr
type of the model.}

\item{\code{type}}{\code{formula()}\cr
model formula.}

\item{\code{type}}{\code{character(1)}\cr
terms of the model. This gets generated using \code{stats::terms} on \code{formula}
during initialisation.}
}}

