% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-register.R
\name{pop_register}
\alias{pop_register}
\title{Population register}
\usage{
pop_register(x, ind_data, hh_data = NULL)
}
\arguments{
\item{x}{an R6 \link{Population} object.}

\item{ind_data}{a data.table that contains individual data}

\item{hh_data}{a data.table that contains household data}
}
\value{
a list contains ind_data and hh_data as data.tables.
}
\description{
Assign new ids to new population data.
}
\note{
It is expected that the provided data in \code{ind_data} and \code{hh_data} have the same
structure or columns as the individual and household data in the \link{Population} object
in \code{x}. Also, it is crucial that all ids in the relationship columns do exist in
the data of the new population data.
}
\examples{

create_toy_world()

pop <- world$get("Population")

new_population_data <- pop_register(pop, toy_individuals, toy_households)
}
