% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{Population}
\alias{Population}
\title{Population}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[dymiumCore:Container]{dymiumCore::Container}.}
\description{
Population contains an instance of \link{Individual} and \link{Household} and methods
that modify both if its contained objects simutaneously.
}
\section{Construction}{
\preformatted{Pop <- Population$new()
}
\itemize{
\item ind_data::\link[data.table:data.table]{data.table::data.table}\cr
Microdata of Individuals/Persons.
\item hh_data::\link[data.table:data.table]{data.table::data.table}\cr
Microdata of Households.
\item pid_col::\code{character(1)}\cr
Individual/Person id column in \code{ind_data}.
\item hid_col::\code{character(1)}\cr
Hoysehold id column in \code{hh_data}
}
}

\section{Public Fields}{

\itemize{
\item \code{ind}:: \code{NULL} | an \link[R6:R6Class]{R6::R6Class} object\cr
Shorthand to the \link{Individual} object.
\item \code{hh}:: \code{NULL} | an \link[R6:R6Class]{R6::R6Class} object\cr
Shorthand to the \link{Household} object.
}
}

\section{Public Methods}{

\itemize{
\item \code{initialise_data(ind_data, hh_data, pid_col, hid_col)}\cr
Load the microdata of individuals and households to construct \link{Individual} and
\link{Household} objects.
\item \code{add_population(ind_data, hh_data)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}, \code{\link[data.table:data.table]{data.table::data.table()}})\cr
add a new population. This requires that all individuals \code{ind_data} belong
to valid households. In the case, that \code{hh_data} is not provided, household ids of
\code{ind_data} will be checked against household ids of the existing households inside
the Population object being added to. One may use \code{pop_register()} to replace
the id columns of the new population with ids that don't exist in the Population object.
\item \code{join_household(ind_ids, hh_ids)}\cr
(\code{integer()}, \code{integer()})\cr
Individuals join their new households and the households' affected attributes,
from the joining of new members, will also be updated. Note! All individuals
must not be in any household prior to joining a new one. Individual's
existing household can be removed using \code{leave_household}.
\item \code{leave_household(ind_ids)}\cr
(\code{integer()})\cr
Remove the individuals' household ids in ind_ids and update the households'
affected attributes, from members leaving the household.
\item \code{remove_emptied_households()}\cr
Remove emptied households.
\item \code{remove_population(pid = NULL, hid = NULL)}\cr
(\code{integer()}, \code{integer()})\cr
Remove population from \verb{$ind} and \verb{$hh} of this \code{Pop} object. If only \code{hid}
is given all household members of households in \code{hid} arg will be removed.
To remove only individuals leave \code{hid} to NULL and specify individuals by their ids
in \code{pid}.
\item \code{inspect(ind_ids = NULL, see_hh = FALSE)}\cr
(\code{integer()}, \code{logical(1)})\cr
Print to console data of individuals and their households.
\item \code{inspect_members(hhid = NULL)}\cr
(\code{integer()}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns Individual$data of all members in \code{hhid}.
\item \code{count_all(verbose = TRUE)}\cr
Print out the number of individuals and households to console.
\item \code{get_hhsize(hids)}\cr
(\code{integer()})\cr
Get hhsize from individual's data and merge it to household data.
\item \code{update_hhsize()}\cr
Update household size.
\item \code{update()}\cr
mask all the household update functions that need to be adjust after changes
in household members or in their attributes; such as change in partnership status,
change of income, birth.
\item \verb{keep_log(var, value, time = .get_sim_time()}\cr
(\code{character(1)},\code{list}|\code{integer(1)}|\code{character(1)}|\code{logical(1)}, \code{integer(1)})\cr
Keep log of events. \code{var} usually uses one of these prefixes "all", "occ", "count" or "id".
\item \code{get_log(type = "all")}\cr
(c("all", "occ", "count", "id")) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Get the event log stored by \code{keep_log}.
\item \code{check_unique_id_cols(ind_data, hh_data = NULL)}\cr
(\code{\link[data.table:data.table]{data.table::data.table()}}, \code{\link[data.table:data.table]{data.table::data.table()}}) -> \code{logical(1)}\cr
Check that all id cols of the input data are unique from the existing ids in
their respective objects.
}
}

