% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scenario.R
\name{set_active_scenario}
\alias{set_active_scenario}
\title{Set an active scenario directory.}
\usage{
set_active_scenario(name)
}
\arguments{
\item{name}{name of the scenario folder to become active.}
}
\value{
a list contains scenario directories.
}
\description{
This is useful for event functions to access the current active scenario directory.
}
\examples{

if (FALSE) {
 use_module("test")
 set_active_scenario("test")
}
}
