% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_functions.R
\name{setup_R}
\alias{setup_R}
\title{Setup R}
\usage{
setup_R()
}
\description{
Load common packages and remove stringsAsFactors.
}
\details{
Silently loads packages from core_packages() and bioc_packages(), deactivates stringsAsFactors and gets bioconductor source URL for package manager
}
\author{
Malte Thodberg
}
\seealso{
\code{\link{create_Rprofile}} \code{\link{core_packages}} \code{\link{bioc_packages}}
}

