% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choice.R
\name{simulate_choice}
\alias{simulate_choice}
\alias{simulate_choice.train}
\alias{simulate_choice.glm}
\alias{simulate_choice.data.frame}
\title{Simulate a choice situation}
\usage{
simulate_choice(model, ...)

\method{simulate_choice}{train}(model, newdata, target = NULL, ...)

\method{simulate_choice}{glm}(model, newdata, target = NULL, ...)

\method{simulate_choice}{data.frame}(model, target = NULL, ...)
}
\arguments{
\item{model}{a \link{Model} object or an object in \code{\link[=SupportedTransitionModels]{SupportedTransitionModels()}}.}

\item{...}{dots}

\item{newdata}{a data.frame object to use for making prediction.}

\item{target}{a \link{Target} object or a named list this is for aligning the simulation
outcome to an external target.}
}
\value{
a character vector
}
\description{
This function simulates a choice situation using a model object and data to
predict probabilities.
}
