% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{SupportedTransitionModels}
\alias{SupportedTransitionModels}
\alias{get_supported_models}
\title{Get all object classes that are supported by Transition}
\usage{
SupportedTransitionModels()

get_supported_models()
}
\value{
a character vector
}
\description{
Currently, these classes are supported in the \code{model} argument of the Transition's
constructor:
\itemize{
\item \link[caret:train]{caret::train},
\item \link[stats:lm]{stats::lm},
\item \link[stats:glm]{stats::glm},
\item a numeric vector,
\item a named \code{list}, and
\item \link[data.table:data.table]{data.table::data.table}.
}
}
\note{
See the 'Transition' section of the
\href{https://core.dymium.org/articles/dymium-intro.html}{dymiumCore's introduction}
webpage for more detail.
}
