% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target.R
\name{Target}
\alias{Target}
\alias{Models}
\title{Target}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting \link{Generic}.}
\description{
\code{Target} is to be used within \code{TransitionClassification} or supply to event
functions. If the target is dynamic then its \code{get} will return its target
value at the current time or its closest time to the current time.
}
\section{Construction}{
\preformatted{Target$new(x)
}
\itemize{
\item \code{x} :: any object that passes \code{check_target()}\cr
A target object.
}
}

\section{Active Field (read-only)}{

\itemize{
\item \code{data}:: a target object\cr
A target object.
\item \code{dynamic}:: \code{logical(1)}\cr
A logical flag which indicates whether the target object is dynamic or not.
}
}

\section{Public Methods}{

\itemize{
\item \code{get(time = .get_sim_time())}\cr
(\code{integer(1)}) -> a named \code{list()}\cr
Get a alignment target as a named list.
}
}

\examples{

TrgtStatic <- Target$new(list(yes = 10))
TrgtStatic$data
TrgtStatic$dynamic
TrgtStatic$get()

target_dynamic <- data.table(time = 1:10, yes = 1:10)
TrgtDynamic <- Target$new(list(yes = 10))
TrgtDynamic$data
TrgtDynamic$dynamic
TrgtDynamic$get()
TrgtDynamic$get(1)
TrgtDynamic$get(10)
}
