% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{trans}
\alias{trans}
\title{Simulate a transition of entities}
\usage{
trans(entity, model, target = NULL, targeted_agents = NULL, update_attr = NULL)
}
\arguments{
\item{entity}{an \link{Entity} object}

\item{model}{a model object that belongs to the classes in \link{SupportedTransitionModels}.}

\item{target}{a named list that is the target for alignment.}

\item{targeted_agents}{a integer vector that contains ids of \code{entity} to undergo
the transition.}

\item{update_attr}{default as NULL. This indicates whether \code{entity} should be updated
using the outcomes from the transtion. To update an attribute of \code{entity} the name
of the attribute to be updated must be specified as a character value.}
}
\value{
a data.table with two columns: id and response.
}
\description{
This function warps \link{TransitionClassification} and \link{TransitionRegression}. It
figures out what is the type of the given model.
}
\examples{

# load data
create_toy_population()
Ind <- pop$get("Individual")

# fit a OLS regression model
model_lm <- glm(age ~ sex + marital_status,
                data = Ind$get_data(),
                family = "gaussian")

# fit a logit model
model_glm <- glm(I(sex == "male") ~ age + marital_status,
                 data = Ind$get_data(),
                 family = "binomial")

# simulation transitions
trans(Ind, model_lm)
trans(Ind, model_glm)
}
