% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionClassification.R
\name{TransitionClassification}
\alias{TransitionClassification}
\title{Transition for classification model}
\arguments{
\item{x}{an \link{Entity} object}

\item{model}{any objects of type in \link{SupportedTransitionModels}.}

\item{target}{a named list where the names corresponds to the choices and the values
are the number of agents to choose those choices. This imposes an alignment of
the outcomes to an external constraint.}

\item{targeted_agents}{a integer vector that contains ids of the entities in \code{x}
to undergo this}

\item{model_by_id}{see in the public field section.}
}
\value{
an \link[R6:R6Class]{R6::R6Class} object
}
\description{
TransitionClassification

Create a \link{TransitionClassification} object.
}
\section{Fields}{

\describe{
\item{\code{model_by_id}}{(\code{logical(1)}) is default as \code{FALSE}. This flag is to indicate
whether the \code{model} object is meant to be matched by the id column of the entity object
in \code{x} or not. It should be noted that this flag only matters if the \code{model} object is
of type \code{\link[data.table:data.table]{data.table::data.table()}} where it must contains a numeric column called \code{prob}
or list columns of type numeric and character called \code{probs} and \code{choices}. The model
object must have a column which its name matches with the id column of the entity object in \code{x}.}
}}

\examples{

# create a Individual agent object
Ind <- Individual$new(.data = toy_individuals, id_col = "pid")

# create a probabilistic model
driver_status_rate <- data.table::data.table(
  sex = c('male', 'female'),
  probs = list(c(0.3,0.7), c(0.4,0.6)),
  choices = list(c('can drive', 'cannot drive'), c('can drive', 'cannot drive'))
)

# create a Transition for driver status
TransitionCandrive <- R6::R6Class(
  classname = "TransitionCandrive",
  inherit = TransitionClassification
)

TransCanDrive <- TransitionCandrive$new(x = Ind, model = driver_status_rate)

barplot(
  table(TransCanDrive$get_result()[['response']]),
  main = "Transition result: driver status",
  col = c('steelblue', 'salmon')
)
}
