% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unnest_datatable}
\alias{unnest_datatable}
\title{unnest_datatable}
\usage{
unnest_datatable(dt, by_col)
}
\arguments{
\item{dt}{data.table object}

\item{by_col}{reference column}
}
\description{
unnest data.table object
}
